### 2019-03-15 #1324 @whedon
[PRE REVIEW]: Stripy: A Python module for (constrained) triangulation in Cartesian coordinates and on a sphere
**Submitting author:** @lmoresi (<a href="http://orcid.org/0000-0003-3685-174X">Louis Moresi</a>)
**Repository:** <a href="https://github.com/underworldcode/stripy" target ="_blank">https://github.com/underworldcode/stripy</a>
**Version:** 0.7.0
