### 2019-03-15 #1324 @arfon
@lmoresi - please take a look at the errors above with your bibtex file. If you make fixes in your `paper` branch you can recompile with `@whedon generate pdf from branch paper`
