### 2019-03-15 #1324 @arfon
@openjournals/joss-eics - the submitting author suggested `@lheagy` as the handling editor.
