### 2019-03-16 891c01b
updated bibliography - the submission system found some bugs
