+++ paper/paper.bib
@@ -4,14 +4,14 @@ title={The structural evolution of the deep continental lithosphere},
 doi={10.1016/j.tecto.2016.12.004},
 journal={Tectonophysics},
 volume={695},
-pages={81-89}
+pages={81-89},
 year={2016}
 }
 
 @article{MullerEtAl2008,
 author={Müller, R.D. and Sdrolias, M. and Gaina, C. and Roest, W.R.},
 title={Age spreading rates and spreading asymmetry of the world's ocean crust},
-journal={Geochemistry, Geophysics, Geosystems}
+journal={Geochemistry, Geophysics, Geosystems},
 doi={ doi:10.1029/2007GC001743},
 volume={9},
 pages={Q04006},
