### 2019-03-16 7cb6003
Small bugs found in preprint proof
