### 2019-03-16 #1324 @lmoresi
I fixed a couple of (unimportant) bugs in the paper that were obvious once I looked at the pdf ... not sure if whedon is listening to me any more though.
