+++ paper/paper.bib
@@ -12,10 +12,10 @@ year={2016}
 author={Müller, R.D. and Sdrolias, M. and Gaina, C. and Roest, W.R.},
 title={Age spreading rates and spreading asymmetry of the world's ocean crust},
 journal={Geochemistry, Geophysics, Geosystems},
-doi={ doi:10.1029/2007GC001743},
+doi={10.1029/2007GC001743},
 volume={9},
 pages={Q04006},
-year={1008}
+year={2008}
 }
 
 @article{Renka1997a,
