### 2019-03-23 65d10bd
Improve the performance of `lonlat2xyz` and, indirectly, six methods of `sTriangulation`.  This test attempts to illustrate the difference: ` stripack_setup=''' from stripy.spherical_meshes import icosahedral_mesh from stripy.spherical import _stripack mesh = icosahedral_mesh(refinement_levels=4) lons,lats = mesh.lons, mesh.lats '''
numpy_setup='''from stripy.spherical_meshes import icosahedral_mesh
import numpy as np
mesh = icosahedral_mesh(refinement_levels=4)
lons,lats = mesh.lons, mesh.lats
'''

numpy_nolocal ='''
x = np.cos(lats) * np.cos(lons)
y = np.cos(lats) * np.sin(lons)
z = np.sin(l
[…]
