### 2019-04-05 #1324 @lmoresi
It's fine. I just wanted to make sure there were no privacy issues. We have the following suggestions that are mainly from the Earth Science side as we think that is the primary audience for stripy and maybe that will help. We also suggest contacting professor Renka whose original code we wrapped as the starting point for the package (email details below)

  - Paul Wessel (U. Hawaii) - pwessel_at_hawaii.edu - PaulWessel on github
  - Simon Williams (Northwest University, Xi’an, China) - contact details TBA
  - Thorsten Becker (UT Austin) -  twb_at_ig.utexas.edu  - thwbecker on github
  - Robert Renka (U. North Texas ) - robert.renka_at_unt.edu
