### 2019-04-24 #1410 @santisoler
@VivianePons seems like `whedon` is having some problems generating the PDF.
Is this related to the fact that the repository ulr is pointing to the `paper` branch instead of pointing to the repo itself (https://github.com/underworldcode/stripy/)?
