### 2019-04-24 #1410 @santisoler
Hi @lmoresi and @brmather. I was assigned as reviewer for your submitted paper.
Let me read through your PDF, explore your `stripy` repo and experiment with your software.
I'll try to meet the two weeks deadline, but in case I get delayed I'll let you know so you don't get worried about the status of the review.

I'm still thinking my review strategy, but I'll probably will open issues on your repo regarding each section of the checklist (e.g. General Checks, Functionality, Documentation, Software Paper).
Besides that, I'm open to help you solve the issues that may appear. My goal is to contribute to get this published! So don't hesitate asking for help or guidance, I'm willing to help with everything that's on my reach.

I'm sure this will be a nice experience for all of us. Be patient, I'll contact you when I have some news.
