### 2019-04-24 #1410 @lmoresi
Thanks - we’ll work with the way you prefer to review. issues sounds like the obvious strategy. We have a “paper” branch of the code to make sure that any developments we make during the review process do not cause problems for you. Any suggestions you make st this stage will be merged back when the process is done.

Prof Louis Moresi

louis.moresi@unimelb.edu.au<mailto:louis.moresi@unimelb.edu.au>

(w) +61 3 8344 1217

(m) +61 4 0333 1413

(us) +1 505 349 4425

www.moresi.info<http://www.moresi.info/>

www.facebook.com/underworldcode<http://www.facebook.com/underworldcode>

@LouisMoresi<https://twitter.com/LouisMoresi>
