### 2019-04-25 #1410 @santisoler
@lmoresi @brmather

> We have a “paper” branch of the code to make sure that any developments we make during the review process do not cause problems for you. Any suggestions you make st this stage will be merged back when the process is done.

Thanks for pointing this out. I've started reviewing some of the more basic aspects of the publication.
I'll open some issues on your repo so we can start talking about them.
I'll try to get into the other subjects next week, so I can continue opening issues if they are needed.
So far the paper is looking great and most of the issues I found are really minor and very easy to tackle down.
