### 2019-04-25 underworldcode/stripy#14 @santisoler
[REVIEW] References on paper
Part of https://github.com/openjournals/joss-reviews/issues/1410

The `paper.md` satisfies all JOSS requirements, it's very well written and the
*statement of need* is very clear within the first two paragraphs.
Nice work! :+1:

I have just one doubt. Does the *Laske, G., Masters, G., Ma, Z., & Pasyanos, M.
(2013)* article has a DOI? I haven't been able to find it through a quick search, but if you
have it, would be nice to add it.

JOSS requires that referenced papers, datasets and softwares must have a DOI, but don't
know what's the policy about abstracts.
In case you can't find a DOI for *Laske et al. (2013)* we could ask the editors what
should we do.
