### 2019-04-25 underworldcode/stripy#15 @santisoler
[REVIEW] License
I see you have chosen the LGPL v3 license for your software. Great choice! :+1:
I think the license is well applied according to the **gnu.org** requirements.
You added:
- the `COPYING` file,
- the `COPYING.LESSER` file,
- license notes to each code file.

I would only change the [`License.txt`](https://github.com/underworldcode/stripy/blob/paper/Licence.txt) file for the one that GitHub offers you to specify the software is released under LGPL v3. If you do it, GitHub will recognize the license and add it to the metadata of the repo. This makes easier to search software by license inside GitHub.

See how Github shows the license of [verde](www.github.com/fatiando/verde) and [yellowbrick](https://github.com/DistrictDataLabs/yellowbrick) on the top bar along with **Commits**, **Branches**, **Releases**.

![screen-verde-license](https://user-images.githubusercontent.com/11541317/56760285-40583a00-6771-11e9-9195-c7350064a22e.jpg)
![screen-yellowbrick-license](https://user-images.githubusercontent.com/11541317/56760459-94631e80-6771-11e9-935e-25fa156404be.jpg)

You can read [this article](https://help.github.com/en/articles/adding-a-license-to-a-repository) for more instructions on how to add a `LICENSE` file in a quick way.

If you have any doubts, don't hesitate to ask for help.
