### 2019-04-25 underworldcode/stripy#16 @santisoler
[REVIEW] Branching and Versioning
Having a dedicated branch for the submitted paper to JOSS is a good idea because current developments won't interfere with the ongoing review, as you pointed out on the [joss-review issue](https://github.com/openjournals/joss-reviews/issues/1410).

Nevertheless I see a few minor issues regarding this approach.
On the first place, the default branch (`master` branch in this case) is like the presentation card of your repo.
It's not common that a user must change branch in order to get a specific version of the software or get some important information about it (like the paper for instance).
Secondly, I noticed two issues regarding your versioning:
1. The `v0.7.0b` release points to a35c04e42ea278a59383509c1ed6cf94e9418b4f on `master` branch, what is not in
   agreement with the `paper` branch that you request for review.
   Besides, the releases are usually generated from commits on the default branch
   (`master`). A release generated from a non-default branch is not
   something a user would expect.
2. The submitted `0.7.0` release does not exists on the repo. A `v0.7.0b` does, but
   I understand that the submitted version must be a valid GitHub release.

I can think an easy solution to this issues. I don't want to interfere with your git
workflow, so take this instructions just as proposals.
1. Create a `dev` branch from the current `master`, where you can continue your
   development if you don't want to interfere with the review.
2. Merge the `paper` branch into `master`.
3. Create a new release (`v0.7.1` for example) with the latest commit on the new
   `master` branch.
4. Then, you could ask @VivianePons to change the version of the submission in order to be
in agreement with the one I'll be reviewing.

Please, feel free to disagree if you don't like this approach.
We can discuss this issues and create better solutions if you feel it's needed.
Also, I'm glad to help you if you have any doubts when applying these solutions.
