### 2019-04-25 underworldcode/stripy#17 @santisoler
[REVIEW] Installation
I really liked the installation instructions: they offers different ways to get `stripy`
working, which is an easy way to encourage users to use it.
I'm more familiar with `pip` package manager than `docker`, although many users would be
very thankful to have this option in case they don't want to bother installing
`gfortran` or having to reconfigure the build if they are using another Fortran
compiler.

I think the **installation instructions requirements of JOSS are already satisfied**,
although I would like to propose some minor changes, specially with
the non-`docker` way.

1. I would specify which minimum Python version is needed. Nowdays everyone is leaving
   Python 2.7 behind because its End of Life is less than a year ahead, so one would
   expect `stripy` to run under Python 3, but would be nice to make it clearer,
   e.g. *You'll need Python 3.5 or greater.*
2. The dependencies can be specified using markdown lists instead of having them on
   a plain sentence.
   For example:
   ```
   ### Dependencies
   - gfortran
   - numpy
   ```
3. The `docker` and the non-`docker` instructions can be included under the same
   **Installation** section, creating a subsection for each installation method.
   So, one could create a subsection called **Installing with pip**, another one like
   **Installing the latest development version** and **Installing with Docker**.
   See [Verde's installation instructions](https://www.fatiando.org/verde/latest/install.html) for inspiration.
4. I think the `python setup.py build && python setup.py install` instructions are
   not the best choice for end users. In my opinion is better to work with `pip` because
   it have great tools to install, uninstall or update your package.
   So I would just leave the `pip` instructions for the sake of convenience.
   If there's any reason why you think you should keep the `python setup.py ...`
   instructions, please let me know.

I'll open a Pull Request applying this modifications so you can then review the changes
and we can discuss there if you agree with them.

## Minor versioning issue

There's a minor issue regarding the release version of the installed package.
If I install `stripy` through `pip` by downloading the latest version on `pypi.org`:
```
pip install stripy
```
I get the `0.7.0` version.
But, if I download the latest version, say the `HEAD` of `paper` branch, it's signed as
the `0.7.0` version although there could be differences with the proper `0.7.0` version.

The problem raises because the `version` field on `setup.py` is hardcoded to
`0.7.0`.
One way to solve this is to hardcode the `version` field to something like `0.7.0-dev`
for non releases commits, and change it only on commits that will be released according
to its version number.
Another way is to automatize the value of the `version` field.
For example, [`verde`](https://www.github.com/fatiando/verde) sets the version based on the hash of the current commit,
which comes very handy when debugging.

This is a good feature to add, but **won't be required to submit the paper to JOSS**.
I just wanted to pointed it out in case you're interested.
If you want, we could leave this as final step before publication, or even after it gets
published.
I can help you implementing it.
