### 2019-04-25 underworldcode/stripy#19 @santisoler
[REVIEW] Error when creating bash_utils.sh from Docker
After installing `stripy` using Docker, I'm following the installation instructions for creating and installing the helper scripts for bash.
So, with the `underworldcode/stripy:latest` container already pulled, I run:
```
docker run --rm underworldcode/stripy:latest bash_utils > bash_utils.sh
```
But, the `bash_utils.sh` contains only the following line:
```
/usr/bin/xvfb-run: line 184: bash_utils: not found
```

Do you have any idea what could be failing?

**System specifications:**
- OS: Manjaro Linux XFCE 64bits
- Docker version 18.09.5-ce, build e8ff056dbc
