### 2019-04-25 underworldcode/stripy#19 @lmoresi
Hmmm ... we may be missing something in the tool chain that gets loaded up when building the docker image. I’ll need to check that again.

Prof Louis Moresi

louis.moresi@unimelb.edu.au<mailto:louis.moresi@unimelb.edu.au>

(w) +61 3 8344 1217

(m) +61 4 0333 1413

(us) +1 505 349 4425

www.moresi.info<http://www.moresi.info/>

www.facebook.com/underworldcode<http://www.facebook.com/underworldcode>

@LouisMoresi<https://twitter.com/LouisMoresi>
