### 2019-04-26 underworldcode/stripy#19 @lmoresi
Ok a it’s a missing package from the dockerfile that must have got lost when we switched to a different base image. I’ll fix it up. How does this work with the submission / release ?

Prof Louis Moresi

louis.moresi@unimelb.edu.au<mailto:louis.moresi@unimelb.edu.au>

(w) +61 3 8344 1217

(m) +61 4 0333 1413

(us) +1 505 349 4425

www.moresi.info<http://www.moresi.info/>

www.facebook.com/underworldcode<http://www.facebook.com/underworldcode>

@LouisMoresi<https://twitter.com/LouisMoresi>
