### 2019-04-26 underworldcode/stripy#19 @santisoler
> Ok a it’s a missing package from the dockerfile that must have got lost when we switched to a different base image. I’ll fix it up. How does this work with the submission / release ?

So it wasn't a big issue, great! :+1:
If you have to make a change on the repo in order to fix the Docker, you could start a Pull Request with the fixing and then merge it into the `paper` branch.
This way we can keep track of which changes fixes which issues (if you write `Fixes #...` on the PR description, the issue gets closed when the PR is merged - see how I referenced #17 on #18).
If you don't know how to do it, you could just commit the changes into the `paper` branch, no big deal.

I'm thinking we could apply future changes to the current `paper` branch and after acceptance you could make a new release. I'll ask the editors for guidance on this subject.
