### 2019-04-26 #1410 @santisoler
Hi @VivianePons. I have a small question regarding how versioning works on the review process.
The authors have submitted the `0.7.0` version of `stripy`, although on the review process they'll surely make some changes.
Is it possible that after acceptance the authors create a new release, `0.7.1` for instance, and make that the *final* version?
Thanks in advance!
