### 2019-04-26 underworldcode/stripy#19 @lmoresi
So, I have updated the docker image. It's not always obvious how to track versioning through this process. The changes are reflected in the dockerfile that we (happen to) keep in our repository. I pushed this to the master branch and merged into the paper branch. To test the docker help you will need to re-pull the docker image. As we haven't talked much about the versioning for this, I haven't changed the docker tag or the README.md in the repo ... happy to take advice on the best way to reflect these kinds of changes in the manuscript versioning
