### 2019-04-26 underworldcode/stripy#19 @santisoler
> So, I have updated the docker image. It's not always obvious how to track versioning through this process. The changes are reflected in the dockerfile that we (happen to) keep in our repository. I pushed this to the master branch and merged into the paper branch. To test the docker help you will need to re-pull the docker image.

That's great @lmoresi, that should do the trick.
Perhaps is cleaner and simpler to make this changes directly on the `paper` branch, so you don't clutter it with the latest changes in `master`.
Nevertheless, I think this issues are related with how your git workflow has been defined (see #16 for my review about it). There's no big problems regarding it, but I think it will eventually give you some headaches.

I've just pulled the new Docker image.
The `bash_utils.sh` is now correctly created, although I'm getting the following error when I run `stripy-docker-browse` after sourcing the bash script.
```
Navigate to http://localhost:8899 to view the examples, ^\ when done
"docker run" requires at least 1 argument.
See 'docker run --help'.

Usage:  docker run [OPTIONS] IMAGE [COMMAND] [ARG...]

Run a command in a new container
```
Seems that the Docker name is missing on the command lines inside `bash_utils.sh`. Am I right?
