### 2019-04-26 underworldcode/stripy#19 @santisoler
> As we haven't talked much about the versioning for this, I haven't changed the docker tag or the README.md in the repo ... happy to take advice on the best way to reflect these kinds of changes in the manuscript versioning

I'm not familiar with tagging Docker containers, but seems that following the same version number than the one assigned to GitHub Releases is the best choice.
Nevertheless, for cases like this where you change the Docker container but don't make a GitHub Release,  you could use the hash of the commit for versioning your Docker. This allows you to keep track of which commit has created that Docker container. [This article](https://container-solutions.com/tagging-docker-images-the-right-way/) recommends this approach and gives more insight on how to do it.

I'm still thinking we should not care too much about versioning on the review process. We should just apply the changes on the `paper` branch (or on the `master` branch if you agree with #16) by merging Pull Requests or just by pushing commits to it.
After acceptance you could create a new release and then tag it as the *final* reviewed version.
I'm still waiting for guidance from the editors on this subject, but don't worry too much about it for now.
