### 2019-04-26 underworldcode/stripy#19 @lmoresi
I'm not sure - I wonder if it is my inadequate understanding of bash functions. I tidied up a little bit and re-enabled the command-line killing option which had stopped working. If you can `docker pull` again, would be great.

I have this as the bash function that gets defined after the templating process spits out the bash_utils.sh. My version of bash is` GNU bash, version 3.2.57(1)-release (x86_64-apple-darwin18)
Copyright (C) 2007 Free Software Foundation, Inc`

```
# Open the default version of the docker to browse the examples
stripy-docker-browse(){
  PORT=${1:-8899} ;
  echo "Navigate to http://localhost:${PORT} to view the examples, ^\ when done" ;
  docker run -it --rm -p ${PORT}:8888 -v ${PWD}:/home/jovyan/external underworldcode/stripy:0.7 ;
}
```
