### 2019-04-26 underworldcode/stripy#19 @santisoler
I'm not used to working with Docker, so my lack of experience didn't let me see an obvious mistake.
That's why I limited myself to run the instructions on the `README.md`.

I've just realized that the `docker pull underworldcode/stripy:latest` didn't pull the `underworldcode/stripy:0.7` version. I thought the `latest` version was just a pointer to the newest tag, like GitHub's `latest` version works. My bad, sorry for the inconvenience.

Now `bash_utils.sh` matches the one on `paper` branch: [`paper:Docker/scripts/bash_utils.sh`](https://github.com/underworldcode/stripy/blob/master/Docker/scripts/bash_utils.sh).

I think this issue is solved, so feel free to close it when you're ready.
Thanks for helping me to solve this.
