### 2019-04-28 underworldcode/stripy#19 @lmoresi
The "latest" tag is the cause of a lot of problems but I suppose we are stuck with that. I'll try to remember that this is a potential issue.

I will close this particular one.
