### 2019-04-29 underworldcode/stripy#14 @lmoresi
The alternative is to cite the web address where the dataset is available or the URL of the abstract PDF. But the Abstract has a large number of citations in the form we have used and no obvious way to find a DOI. I went back to the original location here: https://meetingorganizer.copernicus.org/EGU2013/posters/12567 and https://meetingorganizer.copernicus.org/EGU2013/EGU2013-2658.pdf

Maybe one day the EGU will back-issue DOIs
