### 2019-04-29 underworldcode/stripy#14 @lmoresi
PS - I'll leave this open until there is editorial feedback.
