### 2019-04-29 underworldcode/stripy#15 @lmoresi
Zenodo also likes this better, I believe, so we will take the suggestion on board.
