### 2019-04-29 underworldcode/stripy#17 @lmoresi
I would like to address this issue and, perhaps the one with the docker images having more helpful tagging. It would be best to have tags sync'd between docker and pypi.

How does commit-based tagging work with pip where there is an implied ordering of the tags to make >= operations meaningful ?

Happy to address this offline if it will blow out the review thread.
