### 2019-04-29 d87ca53
Updated dockerfile with fixed base image version
