### 2019-04-29 underworldcode/stripy#16 @lmoresi
OK - so the idea of the paper branch was that, once this process is complete, we would like to have a binder-addressable branch with a dockerfile that people can use to reproduce / rerun the paper for the relevant release of the code. We thought it would also be useful in the review process but probably overlooked the release requirements of joss in setting that up.

This is slightly different from the idea of tagging and releasing an immutable snapshot. I have a similar workflow in a data-interpretation paper in which there is a binder version of the workflow available in a persistent branch that reflects exactly the version we published. The master branch will also include updates to the model as new data are received and re-processed.

Regardless of that motivation, the workflow you propose is also fine. Any branches we make for binder or a successor product do not have to interfere with the joss workflow and we can do that sort of thing later. At the moment, everything is in sync across both branches.

The only thing that remains to fix is the question of the release ... and there I am confused since we have already merged a couple of your suggestions into the master branch ! Any requests ?
