### 2019-04-29 #1410 @VivianePons
Yes, don't worry about the version. When the paper is approved, we update the version before the final publication
