### 2019-04-29 #1410 @VivianePons
I have a question for @arfon : I have contacted Robert Renka to be a second reviewer on this paper (as suggested by the author). He would be willing to do it but he is also the author of the original Fortran code on which the software is based. He is wondering if this is a case of conflict of interest.

I believe it's ok. I actually think that his opinion on the paper would be very valuable. Besides, we have a second independent reviewer so the publication is not based on his opinion alone. What do you think?
