### 2019-04-29 #1410 @arfon
> I believe it's ok. I actually think that his opinion on the paper would be very valuable. Besides, we have a second independent reviewer so the publication is not based on his opinion alone. What do you think?

👍 this sounds OK to me. Thanks for checking.
