### 2019-04-30 #1410 @santisoler
> Yes, don't worry about the version. When the paper is approved, we update the version before the final publication

Thanks for answering this @VivianePons!
