### 2019-04-30 underworldcode/stripy#14 @santisoler
@VivianePons, we have a minor doubt regarding one of the references. Neither the authors nor me could be able to find a DOI for *Laske, G., Masters, G., Ma, Z., & Pasyanos, M. (2013)*.
Can they keep it without a DOI? Can they just add a link to the PDF version of the article instead?
