### 2019-04-30 underworldcode/stripy#15 @santisoler
Great! Because it's expected the `LICENSE` file to be on the root of the repo, most online tools (like GitHub or Zenodo) will try to read it from that location.

I would recommend to reopen this issue until is fixed so we don't forget to tackle it down.
