### 2019-04-30 underworldcode/stripy#15 @santisoler
Sorry, I've just seen that you fixed this on the `master` branch.
Leave this issue closed then, sorry for the inconvenience.
