### 2019-04-30 underworldcode/stripy#16 @santisoler
Thanks for enlighten me on your git workflow.
I think that because git is a very powerful tool it open the possibilities for very different workflows.
Nevertheless there's one absolute true behind it: the best repo management strategy is the one that best fits your needs.
I'm seeing that the persistent branch approach is not very well suited for your needs.
I think this approach is like using branches as a versioning tool, e.g. having a dedicated branch for binder and another one for the reviewed paper is using branches for pointing to different versions of the same repository.

Branches are not well fitted to serve this purposes, while creating tags as immutable snapshots is a great tool for packaging, change logging, bug tracking and debugging.
This is the most common way to maintain git repositories, so most online tools are built to expect this kind of workflow.
For example, if you want to create a binder based on a specific release of your repo you can do it by specifying it. So for a repo with a single `master` branch you could have one binder version for each release plus a binder for the latest commit of the branch.

I don't have so much experience with Docker, but from what I read you can do something similar with it.
You can create one Docker image for each release, and even images for non-tagged commits with its hash as the Docker tag.

> The only thing that remains to fix is the question of the release ... and there I am confused since we have already merged a couple of your suggestions into the master branch ! Any requests ?

I've seen that you added the `LICENCE` file to the `master` branch instead of adding it to the `paper` branch. So I'm a little bit confused how your `master` and `paper` branch differ.

On these cases I always try to keep everything simpler. I don't see that the `master` branch contains significant changes in comparison with the `paper` branch, so one easy solution is to merge the `paper` branch into the `master` branch. Then you could delete the `paper` branch and continue working on `master`.

So, the simplest workflow is to have a single `master` branch and upload new branches only for opening Pull Requests (or for storing experimental code).
After the paper is accepted you could create a new release, which will be an immutable version of the repo, i.e. the one that has been reviewed on JOSS.

For now, I would not worry about releasing, but I would try to order the branches before applying more modifications.
If you want I can make a PR for this merge, so GitHub can tell us in advance if there's any conflict between them.
