### 2019-04-30 underworldcode/stripy#17 @santisoler
I've just added a comment on #16. It's somehow related to this one, so please read it before the following lines.

> I would like to address this issue and, perhaps the one with the docker images having more helpful tagging. It would be best to have tags sync'd between docker and pypi.

You're absolutely right. All the versions across different services **should** have the same tag.
This will prevent huge headaches!

> How does commit-based tagging work with pip where there is an implied ordering of the tags to make >= operations meaningful ?

Maybe I created some confusion with the hash-based tagging proposal.
You should upload only released versions of your software to pypi, and not using hash-tagged releases.
If any user would want to run a specific version of the software (under a non-tagged commit), they can just download the repository and install it with pip from the local repo.
Moreover, you could run:
```
pip https://github.com/username/my-repo/archive/master.zip
```
It will download the `master` branch and install the software. Which version will you get? The one in the `master`'s `HEAD`.
I don't know if this will work with `stripy` because the `setup.py` is not on the root of the repo, but we can see it in the future. Just wanted to show that `pip` does not need a specific release, it can work on any commit of the repo.

So, you'll not have any problems by using the `>=` syntax if you upload only releases to pypi following a versioning semantic.

On the other hand, Docker images has to be manually created. So what happens if you want to test if a Docker image works as expected but don't want to create a release only to do that? In those cases I would use hash-tagged Docker images. This would make tracking very easy and won't confuse users which Docker version they should run.

A different thing happens with Binder. You could create binders either for releases or for branches, so you could have binders running immutable snapshots (releases) or linked to the latest commit on any branch.
