### 2019-04-30 underworldcode/stripy#17 @santisoler
> Happy to address this offline if it will blow out the review thread.

I think JOSS wants the review process to be open, so I would keep our discussions on Issues and PRs.
I think it's important because it lets editors to monitor if any reviewer is violating the Code of Conduct.

So feel free to fill the issues with a lot of comments!
