### 2019-04-30 underworldcode/stripy#17 @santisoler
Please let me know if I didn't explain myself well on any topic.
