### 2019-04-30 underworldcode/stripy#21 @lmoresi
[REVIEW] python tests are broken / unhelpful
This issue was raised by @rrenka as an installation query and I am moving it here for the formal review process:

> I get various errors from the tests:

```

fib_test.py:  ImportError: No module named stripack

mpas_mesh_test.py:  ImportError: No module named netCDF4

test_Triangulation.py:  TypeError: ev() got an unexpected keyword argument 'dx'

test_sTriangulation.py:  AttributeError: 'sTriangulation' object has no attribute 'gradient'

x1.2562.grid.nc:  SyntaxError: Non-ASCII character '\xa9' in file x1.2562.grid.nc on line 2, but no encoding declared; see http://python.org/dev/peps/pep-0263/ for details

```
