### 2019-05-01 f158557
Updated stripy tests and removed broken scripts
