### 2019-05-01 underworldcode/stripy#21 @brmather
Apologies for these error messages, the test scripts have not been updated in some time. This is because all of the components that were tested have been incorporated in the Jupyter notebooks. Regardless, I have removed some redundant tests and fixed `test_routines.py`, `test_Triangulation.py`, and `test_sTriangulation.py` in f1585571636a092de785befbbd1b28c39547796c commit on the master branch.

The latest docker image with the updated tests can be retrieved using the following command,

```sh
docker pull underworldcode/stripy:latest
```

Thanks @rrenka for bringing this to our attention.
