### 2019-05-02 2841964
Config files for binder should be able to be hidden in a subdirectory according to binder examples
