### 2019-05-02 underworldcode/stripy#16 @lmoresi
I am in the middle of doing that.  I will get rid of the paper branch ASAP. Binder doesn't document their API very well for people like me but I dug through their examples more carefully and discovered that we can put the required dockerfile in a subdirectory. This was why I originally had a separate branch: the need to have a DOCKERFILE at the root level (ugly and incompatible with other parts of the workflow). So, having fixed that, I can implement all of the review suggestions here. I will do so, test and then close.
