### 2019-05-02 underworldcode/stripy#16 @lmoresi
Binder now runs from the master branch and I have created a dev branch for any work we do towards the next release while the JOSS review is under weigh.
