### 2019-05-02 underworldcode/stripy#17 @lmoresi
I have restructured the repo to match your suggestions and (after quite a bit of digging through the binderhub / repo2docker repositories) managed to hide the ugly and incompatible binder configuration options that we previously needed branching to manage. Your advice is helpful here and does actually mean we can automate the container process.

My question regarding the tagging was indeed aimed at pypi and so that's probably all fine after your clarification - the use of pip to grab specific repo commits is also fine for ensuring that we have precise reproducibility.

I think we are good to go on this one. Is there anything else ?
