### 2019-05-02 b7a4a5f
Inadvertent merging of old Dockerfile into root directory
