### 2019-05-02 9268cd6
Bug in setup.py
Also force installation of scipy version >= 0.15.0
