+++ stripy-src/setup.py
@@ -39,16 +39,15 @@ if __name__ == "__main__":
           url               = "https://github.com/underworldcode/stripy",
           version           = "0.7.0",
           description       = "Python interface to TRIPACK and STRIPACK fortran code for triangulation/interpolation in Cartesian coordinates and on a sphere",
-          long_description=long_description,
+          long_description  = long_description,
           long_description_content_type='text/markdown',
           ext_modules       = [ext1, ext2, ext3, ext4],
           packages          = ['stripy'],
-          install_requires = ['numpy'],
+          install_requires  = ['numpy', 'scipy>=0.15.0'],
           package_data      = {'stripy': ['Notebooks/CartesianTriangulations/*ipynb',  # Worked Examples is not currently used
                                           'Notebooks/SphericalTriangulations/*ipynb',
                                           'Notebooks/Data/*'] },
           include_package_data = True,
-          install_requires  = ['numpy', 'scipy'],
           classifiers       = ['Programming Language :: Python :: 2',
                                'Programming Language :: Python :: 2.6',
                                'Programming Language :: Python :: 2.7',
