### 2019-05-02 e53ef20
Adding a short test import / version file in the test scripts directory that might be helpful when trying to sort out a source installation
