### 2019-05-02 7fc253a
Adding a short test import / version file in the test scripts directory that might be helpful when trying to sort out a source installation
