### 2019-05-02 underworldcode/stripy#21 @lmoresi
I added a rather inelegant test file for valid importing of various dependencies culled from notebooks etc. The output might be helpful in dealing with the issues above.
