### 2019-05-02 underworldcode/stripy#16 @santisoler
Great! :+1:
Thanks for taking my proposals into account, hope they make the repo management easier.

Just want to note that having several files on the root is very common (see [verde](https://github.com/fatiando/verde), [simpeg](https://github.com/simpeg/simpeg) or [yellowbrick](https://github.com/districtdatalabs/yellowbrick)). So don't be afraid if it looks ugly.
In fact, I think it's better that every portion of the repo (code, web services, continuous integration) work the simpler as they could rather than having files on non regular locations.
Although, this Dockerfile configuration is working, so we should leave it as it is.
