### 2019-05-02 underworldcode/stripy#17 @santisoler
Thanks for taking my proposals into consideration.
I think this issue is already solved, so I'll close it.
If any other issue comes up regarding the installation instructions, we could open a new issue in the future.
