### 2019-05-03 1243033
Updated dependencies in setup.py
And modified README.md
