+++ stripy-src/setup.py
@@ -44,6 +44,7 @@ if __name__ == "__main__":
           ext_modules       = [ext1, ext2, ext3, ext4],
           packages          = ['stripy'],
           install_requires  = ['numpy', 'scipy>=0.15.0'],
+          python_requires   = '>=2.7, >=3.5',
           package_data      = {'stripy': ['Notebooks/CartesianTriangulations/*ipynb',  # Worked Examples is not currently used
                                           'Notebooks/SphericalTriangulations/*ipynb',
                                           'Notebooks/Data/*'] },
