### 2019-05-03 7fd0132
Replace time.clock with time.time
Shields against DepreciationWarning in the test scripts
(at least until time.time is also superseded)
