### 2019-05-06 underworldcode/stripy#24 @santisoler
[REVIEW] Missing dependencies for Notebooks
Part of https://github.com/openjournals/joss-reviews/issues/1410

When running some of the Notebooks the following dependencies are required:
-  `gdal`
- `cartopy`
- `lavavu`
- `pyproj`

Would be nice to have them as dependencies on the `README.md`.
Because they are not essential in order to run `stripy`, they could be added as `Extra dependencies` or `Recommended packages` with a note detailing that they should be installed for running the Notebooks.

I'm willing to help to solve this.
