### 2019-05-06 underworldcode/stripy#25 @santisoler
[REVIEW] Lavavu examples not showing on binder
Part of https://github.com/openjournals/joss-reviews/issues/1410

The interactive plotting of spherical triangulations using `lavavu` is not working as expected on binder.
I'm getting the following output cell:
![lavavu-binder](https://user-images.githubusercontent.com/11541317/57232469-07cb2280-6ff3-11e9-8618-6f552280cd44.png)

Don't know if it's something wrong with binder itself, or if `lavavu` is intended to work under Jupyter Notebooks. Do the notebooks work well locally on your systems?
