### 2019-05-06 underworldcode/stripy#26 @santisoler
[REVIEW] Presence of litho1pt0 on stripy repository
Part of https://github.com/openjournals/joss-reviews/issues/1410

I'm a little bit confused about the presence of `litho1pt0` source code on `stripy` repo.
I can see the vinculation between both packages, and the importance of showing how `stripy` is a great companion for the Litho1.0 dataset.
Nevertheless, it's not a good practice to include several packages into a single repo.
Having one repository for each package makes maintenance a lot easier and makes very clear which is the goal of each package.
For example, `litho1pt0` needs `stripy`, but `stripy` could be applied to other problems besides the ones that `litho1pt0` tries to solve.
So keeping `stripy` as a separated package would be helpful for users that desire to use it for solving these other problems.

Another approach would be to merge both packages into a single one.
This could be done including the `litho1pt0` source code into the `stripy` code.
Nevertheless, I don't think this is the best solution for your case: it would mislead the goal of `stripy` making it very tied to `litho1pt0` applications rather than the actual Python interface to `TRIPACK` and `STRIPACK`.

If you choose to create separated packages, you can still keep the Notebooks containing Litho 1.0 examples although `litho1pt0` should be installed as a (secondary) dependency (see #24).
The same goes for `stripy` on the `litho1pt0` repo: it should be included as a main dependency for `litho1pt0`.

I'm glad to hear your opinions about this!
