### 2019-05-06 underworldcode/stripy#27 @santisoler
[REVIEW] Notebooks directory inside stripy source code
Part of https://github.com/openjournals/joss-reviews/issues/1410

Leaving the source code of the package and the notebooks that show examples of its applications separated is a good practice in order to keep the repository better organized.
I found that the `stripy` related notebooks are located inside `/stripy-src/stripy/Notebooks` and linked on the `/stripy/Notebooks` directory.
I think that moving the notebooks to the latter path would tidy up the repo a little bit, making easier to keep source code separated from documentation/examples.

Do you agree? Have you included the notebooks on that location for some special reason?
