### 2019-05-06 underworldcode/stripy#28 @santisoler
[REVIEW] Compile documentation with Sphinx
Part of https://github.com/openjournals/joss-reviews/issues/1410

I've found that the `stripy` public functions and classes are very well documented, and together with the Notebooks they give a nice starting point for newcomers.
So I consider the **Example usage** item on the review checklist is already satisfied with the notebooks, while the **Functionality documentation** is covered with the functions and classes docstrings.
Nevertheless, reading documentation from source code files or asking a specific docstring from a Jupyter Notebook can be uncomfortable on the long term.

In sumary, I think both items are **already satisfied**, but they could be improved with a compiled documentation.
Building and uploading the package documentation to a website is a best practice for software development, making a fast a very helpful way to learn how to use its functions and classes.

One of the most used tools to do so is [Sphinx](https://www.sphinx-doc.org/en/stable/).
It can automatically generate an HTML version of your package documentation by automatically compiling your Python functions' docstrings.
So, you just need to configure Sphinx and not worry about copying and pasting documentation.
Besides, an elegant HTML documentation is a very nice start point for newcomers.

A lot of packages use Sphinx for building their documentation. Just for citing some of them, see [verde](https://github.com/fatiando/verde) or [yellowbrick](https://github.com/districtdatalabs/yellowbrick).

The files for building the documentation are usually under the `docs` directory.
After Sphinx is run, it generates the website as HTML files, which could be uploaded to be available online.
This task can be automated by using [Travis CI](https://travis-ci.org/) or any other continuous integration service (Travis CI is completely free of charge for GitHub public repos).
So for example, you could have a `docs` directory on your repo (under the `master` branch), configure Travis CI to build a new version of the website after a new commit on `master` is pushed and then make Travis to deploy the HTML files to a `gh_pages` branch inside the same repo.
Github offers you the option to serve a static website on any specific branch (also for free), so in this case you should choose `gh_pages`.
In conclusion, your documentation will be automatically generated and uploaded to https://underworldcode.github.io/stripy making it publicly available.

If you consider that generating a documentation would be helpful and want to implement it, let me know.
I'm willing to help you to tackle this.
