### 2019-05-06 underworldcode/stripy#30 @santisoler
[REVIEW] Missing community guidelines
Part of https://github.com/openjournals/joss-reviews/issues/1410

JOSS requires that the documentation (and/or the `README.md`) must have guidelines to developers and users on how to contribute to the software, how to report issues and problems with it and how to seek support.

[`MetPy`](https://github.com/Unidata/MetPy/) has a very nice [contributors guide](https://github.com/Unidata/MetPy/blob/master/CONTRIBUTING.md) that you can use for inspiration.
I also like [`verde`'s contributing guide](https://github.com/fatiando/verde/blob/master/CONTRIBUTING.md).
Because they are released under open-source licenses, you are free to copy and paste them, and modify them in case you need to.
