### 2019-05-06 underworldcode/stripy#27 @lmoresi
The original reason was to be able to bundle them standalone with the pypi installation and install notebooks after a pip install. We can probably move those to a different location but we usually do not have a separate repository for this level of documentation. I expect we can engineer for setup.py to reach up and out to get the notebooks from elsewhere in the repo.
