### 2019-05-06 underworldcode/stripy#28 @lmoresi
This is probably a good idea, but I wonder if it may be better to do this after the formal review and ask for your help if you are still able to find the time. I would prefer that we give you credit for that level of assistance that is beyond simply a review.
