### 2019-05-06 underworldcode/stripy#26 @lmoresi
OK - this is a historical artifact. Litho1pt0 was the original code and stripy was broken out of it. I would be happy to remove this and create another repo for litho1pt0 given the direction stripy has now taken. [[ @brmather , we should do all the other matters regarding tests and reorganisation first so that litho1pt0 benefits from all the advice we are receiving in this review ]]
