### 2019-05-06 underworldcode/stripy#25 @lmoresi
lavavu should work in the binder but there is a continual battle between jupyter hub security and the various reverse-proxy strategies we have to follow to keep the lavavu service visible.

@brmather - the known working solution is to use lavavu<1.3 (i.e. the final 1.2.x release) and make sure that binder is pointing to traditional notebook rather than jupyterlab. This is probably something that was not checked when I updated the docker images to fix the issues with the bash helpers.

[[@brmather, I'm assigning this to you as I cannot tackle anything this week. Re-assign to me if you need to]]
