### 2019-05-07 underworldcode/stripy#27 @santisoler
Oh ok! I didn't notice that was the reason for the current location of the notebooks.
So, in light of this new reason let me rethink the issue.

The current location is not harmful, so we could consider it as a possible solution.
I don't quite like to force the `setup.py` to find files outside its directory. You'll need to hack it and create a non very elegant solution.

I can see two more options here:

1. There no special reason why Notebooks should be released under pypi. They could be downloaded from `stripy` once it's already installed. For example, you could create a function that downloads the notebooks from your repo:
    ```python
    import stripy

    stripy.download_notebooks()
    ```
    On @fatiando we have a very nice tool for fetching files from the web called [pooch](https://www.fatiando.org/pooch/). It's very friendly when the files are located on a GitHub repo: you could specify what version of the file you want.

2. An easier solution would be to refactor the organization of the repo. Instead of having all the source code related files inside `stripy-src`, you could move all those files to the root. So the repo would look something like this:
    ```
    ├── binder
    ├── Docker
    ├── litho1pt0-src
    ├── Notebooks
    ├── paper
    ├── stripy
    │   ├── f77-src
    │   ├── src
    │   ├── __init__.py
    │   ├── cartesian_meshes.py
    │   ├── cartesian.py
    │   ├── documentation.py
    │   ├── spherical_meshes.py
    │   └── spherical.py
    ├── tests
    ├── COPYING
    ├── COPYING.LESSER
    ├── LICENCE
    ├── MANIFEST.in
    ├── README.md
    └── setup.py
    ```
    With this organization, the `Notebooks` directory is located on the same level as `setup.py`, so you could easily add them as `package_data` to pack the notebooks on pypi.
    Moreover, with this organization you would enable the installation of the latest version of the repo with pip (as I mentioned on #17):
    ```
    pip install https://github.com/underworldcode/stripy/archive/master.zip
    ```

Both solutions are not mutually excluding, so you could apply both simultaneously if you consider it convenient.
Let me know if you like any of them or if you could think a third possible scenario.

As usual, I'm willing to help you implementing any solution you decide to apply.
