### 2019-05-07 underworldcode/stripy#28 @santisoler
Sounds good to me, we could leave the documentation compilation for the future and focus on more *urgent* issues.

Please, leave this issue open to remind us that we want to tackle this on the future.

PS: Thanks for your consideration about my assistance. I think the open review opens the possibility to create these kind of contributions between people around the world. I'm glad I can help and experience this very nice way to perform peer review.
