### 2019-05-07 underworldcode/stripy#26 @santisoler
:+1: In fact you could submit another paper with `litho1pt0` to JOSS :tada:
