### 2019-05-08 underworldcode/stripy#26 @brmather
Yes, I agree it makes sense to split `litho1pt0` out of stripy. The documentation for it is not up to the same standard as `stripy` so this is something we could work on separately. Thanks @santisoler for the suggestion.
