### 2019-05-08 underworldcode/stripy#27 @brmather
I personally prefer your second proposal. I think the original motivation for keeping everything tucked away in `stripy-src` was to de-clutter the root directory, but in hindsight that has created unforeseen issues where license files have not been detected and symbolic links to README.md were required. If we are moving `litho1pt0` source to another repository (#26) I see little need to store stripy source code in a subdirectory. @lmoresi what do you think?
