### 2019-05-08 underworldcode/stripy#27 @lmoresi
Yes, that is my preference too. The issue with grabbing notebooks from downloads is also going to be one of versioning. As we do with underworld, running the example notebooks should be one of the tests in the test suite so it is useful for that to be something we can do on the repository without needing to synchronise releases in multiple repos.

The proposed structure is fine and I would suggest that we 1) reorganise (including litho1pt0), 2) split out litho1pt0 and 3) remove it from the current stripy.

I don’t think litho1pt0 deserves a publication as it is really just a “thin wrapper” which is discouraged by joss. We can give it a zenodo presence, though.

L

Prof Louis Moresi

louis.moresi@unimelb.edu.au<mailto:louis.moresi@unimelb.edu.au>

(w) +61 3 8344 1217

(m) +61 4 0333 1413

(us) +1 505 349 4425

www.moresi.info<http://www.moresi.info/>

www.facebook.com/underworldcode<http://www.facebook.com/underworldcode>

@LouisMoresi<https://twitter.com/LouisMoresi>
