### 2019-05-08 underworldcode/stripy#25 @lmoresi
I just checked with the binder / dockerfile with lavavu that I am using for a class. The current build does work with binder (which has changed its interface in the last few days) even when using the new default jupyter-lab.

So, that is commit **afa8cd24e6dae7ffbacf9735db8e1affd85fd3df** which we should probably specify explicitly in the dockerfile
