### 2019-05-08 693deb2
Fix bug where lavavu does not work on Binder
