### 2019-05-08 underworldcode/stripy#25 @brmather
OK this has been fixed with 693deb22fe203912ac017dbdb6917313c9c8f53e and should *stay* fixed since we have frozen the version of lavavu.
