### 2019-05-08 underworldcode/stripy#25 @lmoresi
Not necessarily - there are some changes with binder that require particular versions of jupyterhub (I think) and this means we might have to update the dockerfile and the various lavavu tricks to integrate with it.

Fingers crossed it continues to work for a decent amount of time.

Prof Louis Moresi

louis.moresi@unimelb.edu.au<mailto:louis.moresi@unimelb.edu.au>

(w) +61 3 8344 1217

(m) +61 4 0333 1413

(us) +1 505 349 4425

www.moresi.info<http://www.moresi.info/>

www.facebook.com/underworldcode<http://www.facebook.com/underworldcode>

@LouisMoresi<https://twitter.com/LouisMoresi>
