### 2019-05-09 underworldcode/stripy#27 @santisoler
Great! Because this refactor could involve major changes, I would recommend creating a PR to do it.
That way we can test if everything is working as expected before merging it.

> I don’t think litho1pt0 deserves a publication as it is really just a “thin wrapper” which is discouraged by joss. We can give it a zenodo presence, though.

Ok. I didn't know `litho1pt0` is not suited for a JOSS publication. But would be nice if it has its own repo along with a Zenodo doi.
