+++ README.md
@@ -1,1 +1,0 @@
-stripy-src/README.md
