+++ setup.py
@@ -28,9 +28,9 @@ ext1 = Extension(name    = 'stripy._stripack',
 ext2 = Extension(name    = 'stripy._tripack',
                  sources = ['src/tripack.pyf', 'src/tripack.f90'])
 ext3 = Extension(name    = 'stripy._srfpack',
-                 sources = ['src/srfpack.pyf', 'f77-src/srfpack.f'])
+                 sources = ['src/srfpack.pyf', 'src/f77/srfpack.f'])
 ext4 = Extension(name    = 'stripy._ssrfpack',
-                 sources = ['src/ssrfpack.pyf', 'f77-src/ssrfpack.f'])
+                 sources = ['src/ssrfpack.pyf', 'src/f77/ssrfpack.f'])
 
 if __name__ == "__main__":
     setup(name = 'stripy',
@@ -56,5 +56,6 @@ if __name__ == "__main__":
                                'Programming Language :: Python :: 3.3',
                                'Programming Language :: Python :: 3.4',
                                'Programming Language :: Python :: 3.5',
-                               'Programming Language :: Python :: 3.6']
+                               'Programming Language :: Python :: 3.6',
+                               'Programming Language :: Python :: 3.7']
           )
