### 2019-05-14 c735c0a
Updated MANIFEST.in
