+++ MANIFEST.in
@@ -10,6 +10,6 @@ include stripy/spherical.py
 include stripy/spherical_meshes.py
 include stripy/cartesian_meshes.py
 include stripy/documentation.py
-include stripy/Notebooks/SphericalTriangulations/*.ipynb
-include stripy/Notebooks/CartestianTriangulations/*.ipynb
-include stripy/Notebooks/Data/*
+include Notebooks/SphericalTriangulations/*.ipynb
+include Notebooks/CartestianTriangulations/*.ipynb
+include Notebooks/Data/*
