### 2019-05-14 underworldcode/stripy#27 @brmather
Alright, I have created a [restructure](https://github.com/underworldcode/stripy/tree/restructure) branch to deal with the moving everything to the top-level directory with this commit c735c0a38dc5e8c53d33d974cc32fdfb4d069d7e. Changes include:

- `/stripy-src` -> `/`
- `/stripy-src/stripy/Notebooks` -> `/Notebooks`

@santisoler: how do you include the Notebooks as `package_data`? I have the following in `setup.py`:

```python
package_data      = {'stripy': ['Notebooks/CartesianTriangulations/*ipynb',
                                'Notebooks/SphericalTriangulations/*ipynb',
                                'Notebooks/Data/*'] },
```

But they are nowhere to be found under stripy package when it is built with `pip` or `distutils`. Could you shed some light on why this is happening?

Many thanks!
