### 2019-05-14 underworldcode/stripy#27 @brmather
Yes it's in there - but why would you define `package_data` if you didn't want it to be included!? Weird option.
