### 2019-05-14 underworldcode/stripy#27 @santisoler
> Alright, I have created a restructure branch to deal with the moving everything to the top-level directory

Great @brmather! Would you mind opening a Pull Request for the `restructure` branch so we could continue the discussion over there? The Pull Requests interface offers a lot of tools for discussing over the proposed code, like reviews, comments or even changes suggestion from any collaborator.
(In case you don't know how to do it, you could get very nice instructions from [this GitHub article](https://help.github.com/en/articles/creating-a-pull-request))

Instead of waiting for that PR, I'll try to answer here so I don't delay you.

I'm seeing that `setup.py` import two `setup` functions:
```python
from setuptools import setup, find_packages
from numpy.distutils.core import setup, Extension
```
The script will ultimately use only the last one of them, so would be better to remove the first import.
I'm not used to `numpy.distutils.core`, so I'm not so familiar with this specific `setup.py`.
Is there a reason why you should import `setuptools` as well?

According to [this StackOverflow answer](https://stackoverflow.com/a/41896134), if you want to use `setuptools` you need to import it as a whole **before** importing the `numpy.distutils.core` related functions and classes:
```python
import setuptools
from numpy.distutils.core import setup, Extension
```
I'm sure you have more knowledge and experience than me on this subject. Please enlighten me.

> @santisoler: how do you include the Notebooks as package_data?

> But they are nowhere to be found under stripy package when it is built with pip or distutils. Could you shed some light on why this is happening?

Ok. Maybe my idea of leaving the `Notebooks` outside the `stripy` directory is not a good one because the files included on `package_data` must be **inside** the package directory.
This is mainly due to the fact that we want the Notebooks to be distributed inside the pypi package.
So I think the best choice would be to keep the `Notebooks` inside `stripy`.
Nevertheless, I think we could still apply some of the changes of `restructure` branch.

Maybe, we can update the repo structure to look like this:

    ├── binder
    ├── Docker
    ├── litho1pt0-src
    ├── paper
    ├── stripy
    │   ├── f77-src
    │   ├── Notebooks
    │   ├── src
    │   ├── __init__.py
    │   ├── cartesian_meshes.py
    │   ├── cartesian.py
    │   ├── documentation.py
    │   ├── spherical_meshes.py
    │   └── spherical.py
    ├── tests
    ├── COPYING
    ├── COPYING.LESSER
    ├── LICENCE
    ├── MANIFEST.in
    ├── README.md
    └── setup.py

Regarding @lmoresi comment about `include_package_data = True`, the [`setuptools.py` documentation](https://setuptools.readthedocs.io/en/latest/setuptools.html?highlight=include%20files#including-data-files) specify three ways for including data files:
- `include_package_data`
    Accept all data files and directories matched by MANIFEST.in.
- `package_data`
    Specify additional patterns to match files that may or may not be matched by MANIFEST.in or found in source control.
- `exclude_package_data`
    Specify patterns for data files and directories that should not be included when a package is installed, even if they would otherwise have been included due to the use of the preceding options.

So, I think if you choose to use the `include_package_data`, the `package_data` could be omitted. The same goes otherwise: if using `package_data`, the `include_package_data` could be omitted.
In conclusion, if you use `package_data = {"..." = ["..."]}` you don't need to specify `include_package_data`.
I'm familiar with `setuptools` and the second option (using `package_data`), but I'm not entirely sure if `numpy.distutils.core` is really using `setuptools` or just `distutils`.
