### 2019-05-15 b1a135d
Moved Notebooks back inside stripy again
