### 2019-05-15 7fc1571
Updating MANIFEST.in
