### 2019-05-15 underworldcode/stripy#27 @brmather
OK I moved the Notebooks back inside `/stripy` again with b1a135daf2189d03b530d538c68444ee102c2f03.

We need `numpy.distutils.core` to compile fortran sources. As far as I'm aware `setuptools` is not capable of doing that.

In your proposed directory restructure, you have kept `litho1pt0-src` in there. I was under the impression this should be moved to a separate repository under the underworldcode group. Is this still the idea? I've removed it and all associated notebooks in this pull request #31
