### 2019-05-15 underworldcode/stripy#26 @lmoresi
In the proposed restructure, litho1pt0 is gone and I have created the relevant repository so that the history is not lost. I think we can close this issue
