### 2019-05-15 underworldcode/stripy#27 @lmoresi
I think we do move it. Some folks like to have related projects together, some do not. However, this one was always an uncomfortable marriage and I prefer the unencumbered stripy to be honest.

In placing the Notebooks within stripy, I followed a pattern that has been advocated elsewhere for distributing documentation in this form. As we have seen from trying alternatives, no choice is perfect but this one does work and it does have some precedent.

When we get to that point, it seems to be a matter of taste and we can defend that with readers of JOSS and users of the code.
