### 2019-05-15 4978636
Moved testing framework to pytest ... sTriangulation tests continue to fail. We could do with more fine-grained tests too.
