+++ stripy-src/setup.py
@@ -42,9 +42,11 @@ if __name__ == "__main__":
           long_description  = long_description,
           long_description_content_type='text/markdown',
           ext_modules       = [ext1, ext2, ext3, ext4],
-          packages          = ['stripy'],
           install_requires  = ['numpy', 'scipy>=0.15.0'],
           python_requires   = '>=2.7, >=3.5',
+          setup_requires    = ["pytest-runner", 'webdav'],
+          tests_require     = ["pytest", 'webdav'],
+          packages          = ['stripy'],     d
           package_data      = {'stripy': ['Notebooks/CartesianTriangulations/*ipynb',  # Worked Examples is not currently used
                                           'Notebooks/SphericalTriangulations/*ipynb',
                                           'Notebooks/Data/*'] },
@@ -56,5 +58,7 @@ if __name__ == "__main__":
                                'Programming Language :: Python :: 3.3',
                                'Programming Language :: Python :: 3.4',
                                'Programming Language :: Python :: 3.5',
-                               'Programming Language :: Python :: 3.6']
+                               'Programming Language :: Python :: 3.6',
+                               'Programming Language :: Python :: 3.7'
+                               ]
           )
