+++ stripy-src/setup.cfg
@@ -1,0 +1,12 @@
+[aliases]
+test=pytest
+
+[bdist_wheel]
+
+# This flag says that the code is written to work on both Python 2 and Python
+# 3. If at all possible, it is good practice to do this. If you cannot, you
+# will need to generate wheels for each Python version that you support.
+
+universal=1
+[metadata]
+description-file = README.md
