+++ tests/test_routines.py
@@ -7,7 +7,7 @@ from time import time
 # global variables
 permute = False
 max_refinements = 2
-npoints = 100000
+npoints = 10000
 
 
 def time_routine(routine, *args):
@@ -19,71 +19,76 @@ def time_routine(routine, *args):
 """
 Test Triangulation routines
 """
-print("==========\nTriangulation routines\n==========")
-
-# Create some (semi) random points
-np.random.seed(0)
-x = np.random.random(npoints)
-y = np.random.random(npoints)
-
-# interpolation data
-xi = np.random.random(100)
-yi = np.random.random(100)
-rint = np.random.randint(0, len(xi), len(xi))
-
-# triangulation
-tri = stripy.Triangulation(x, y)
-
-for refine in range(max_refinements):
-    print("\nrefinement = {}\n".format(refine))
-    time_routine(tri.__init__, x, y, refine, permute)
-
-    z = np.hypot(tri.x, tri.y)
-
-    time_routine(tri.areas)
-    time_routine(tri.convex_hull)
-    time_routine(tri.gradient, z)
-    time_routine(tri.gradient_local, z, 0)
-    time_routine(tri.interpolate_nearest, xi, yi, z)
-    time_routine(tri.interpolate_linear, xi, yi, z)
-    time_routine(tri.interpolate_cubic, xi, yi, z)
-    time_routine(tri.nearest_vertex, xi, yi)
-    time_routine(tri.containing_triangle, xi, yi)
-    time_routine(tri.containing_simplex_and_bcc, xi, yi)
 
 
-"""
-Test  sTriangulation routines
-"""
-print("==========\nsTriangulation routines\n==========")
+def test_cartesian_triangulation():
+    # Create some (semi) random points
+    np.random.seed(0)
+    x = np.random.random(npoints)
+    y = np.random.random(npoints)
 
-# Create some (semi) random points
-np.random.seed(0)
-lons = 2.*np.pi*np.random.random(npoints)
-lats = np.arccos(2.*np.random.random(npoints) - 1.) - np.pi/2
+    # interpolation data
+    xi = np.random.random(100)
+    yi = np.random.random(100)
+    rint = np.random.randint(0, len(xi), len(xi))
 
-# interpolation data
-lon = 2.*np.pi*np.random.random(100)
-lat = np.arccos(2.*np.random.random(100) - 1.) - np.pi/2
-rint = np.random.randint(0, len(lon), len(lon))
+    # triangulation
+    tri = stripy.Triangulation(x, y)
+    assert tri.npoints == npoints
+    return(tri, x, y, xi, yi)
 
-# stri = stripy.sTriangulation(lons, lats)
-stri = stripy.spherical_meshes.octahedral_mesh(include_face_points=True, refinement_levels=6)
-stri = stripy.sTriangulation(stri.lons, stri.lats)
 
-for refine in range(max_refinements):
-    print("\nrefinement = {}\n".format(refine))
-    time_routine(stri.__init__, stri.lons, stri.lats, refine, permute)
+def test_cartesian_refinement():
+    tri, x, y, xi, yi = test_cartesian_triangulation()
 
-    Z = np.cos(5.0*stri.lons) * np.sin(2.0*stri.lats)
+    for refine in range(max_refinements):
+        print("\nrefinement = {}\n".format(refine))
+        time_routine(tri.__init__, x, y, refine, permute)
 
-    time_routine(stri.areas)
-    time_routine(stri.gradient_lonlat, Z)
-    time_routine(stri.interpolate_nearest, lon, lat, Z)
-    time_routine(stri.interpolate_linear, lon, lat, Z)
-    time_routine(stri.interpolate_cubic, lon, lat, Z)
[…]
