### 2019-05-15 b0450d4
pytest fixes
