+++ setup.py
@@ -28,9 +28,9 @@ ext1 = Extension(name    = 'stripy._stripack',
 ext2 = Extension(name    = 'stripy._tripack',
                  sources = ['src/tripack.pyf', 'src/tripack.f90'])
 ext3 = Extension(name    = 'stripy._srfpack',
-                 sources = ['src/srfpack.pyf', 'f77-src/srfpack.f'])
+                 sources = ['src/srfpack.pyf', 'src/f77/srfpack.f'])
 ext4 = Extension(name    = 'stripy._ssrfpack',
-                 sources = ['src/ssrfpack.pyf', 'f77-src/ssrfpack.f'])
+                 sources = ['src/ssrfpack.pyf', 'src/f77/ssrfpack.f'])
 
 if __name__ == "__main__":
     setup(name = 'stripy',
@@ -46,7 +46,7 @@ if __name__ == "__main__":
           python_requires   = '>=2.7, >=3.5',
           setup_requires    = ["pytest-runner", 'webdav'],
           tests_require     = ["pytest", 'webdav'],
-          packages          = ['stripy'],     d
+          packages          = ['stripy'],
           package_data      = {'stripy': ['Notebooks/CartesianTriangulations/*ipynb',  # Worked Examples is not currently used
                                           'Notebooks/SphericalTriangulations/*ipynb',
                                           'Notebooks/Data/*'] },
