### 2019-05-15 underworldcode/stripy#29 @lmoresi
Agreed - pytest is great. The current tests are a bit lengthy for the pytest approach but I have  put them into the pytest format. There is some scipy issue at the moment that needs to be sorted out
