### 2019-05-15 underworldcode/stripy#27 @santisoler
> OK I moved the Notebooks back inside `/stripy` again with [b1a135d](https://github.com/underworldcode/stripy/commit/b1a135daf2189d03b530d538c68444ee102c2f03).

:+1:

> We need `numpy.distutils.core` to compile fortran sources. As far as I'm aware `setuptools` is not capable of doing that.

Yes, you're right. But, from what I've read, I understood that importing `setuptools` would enhance the `setup` function from `numpy.distutils.core`. My main concern about this was not functionality, but the presence of two `setup` functions on the same script, what could cause problems in the future if you don't remember exactly which one you must use.

> In your proposed directory restructure, you have kept `litho1pt0-src` in there. I was under the impression this should be moved to a separate repository under the underworldcode group. Is this still the idea? I've removed it and all associated notebooks in this pull request #31

Yes. Sorry for that, maybe I should have remove it from the structure.
My plan was to solve the notebooks location on this issue, leaving the presence of `litho1pt0` for #26.
That's why I left the `litho1pt0` on the proposed structure, just to make the two issues independent of each other.
