### 2019-05-15 underworldcode/stripy#27 @santisoler
> In placing the Notebooks within stripy, I followed a pattern that has been advocated elsewhere for distributing documentation in this form. As we have seen from trying alternatives, no choice is perfect but this one does work and it does have some precedent.
>
> When we get to that point, it seems to be a matter of taste and we can defend that with readers of JOSS and users of the code.

I agree with that solution. Sometimes we open an issue that won't ultimately solve any problem, but make us rethink why we have designed some piece of software as it is and why it's the best solution we can find.
I think this kind of discussion make software more robust.
