### 2019-05-15 underworldcode/stripy#29 @santisoler
:+1: Let me know if you need any help with those issues!
