### 2019-05-15 0d4aafe
Added some tests, fix a couple of issues, exposed a couple more
