### 2019-05-16 98d09b8
First pass at some S-T manifold mesh.
