+++ setup.py
@@ -11,12 +11,45 @@
 ##
 ## (see http://peterdowns.com/posts/first-time-with-pypi.html)
 
-
-
 from setuptools import setup, find_packages
 from numpy.distutils.core import setup, Extension
 from os import path
-import io
+import os, io
+import subprocess
+
+## in development set version to none and ...
+PYPI_VERSION = None
+
+# Return the git revision as a string (from numpy)
+def git_version():
+    def _minimal_ext_cmd(cmd):
+        # construct minimal environment
+        env = {}
+        for k in ['SYSTEMROOT', 'PATH']:
+            v = os.environ.get(k)
+            if v is not None:
+                env[k] = v
+        # LANGUAGE is used on win32
+        env['LANGUAGE'] = 'C'
+        env['LANG'] = 'C'
+        env['LC_ALL'] = 'C'
+        out = subprocess.Popen(cmd, stdout = subprocess.PIPE, env=env).communicate()[0]
+        return out
+
+    try:
+        out = _minimal_ext_cmd(['git', 'rev-parse', '--short', 'HEAD'])
+        GIT_REVISION = out.strip().decode('ascii')
+    except OSError:
+        GIT_REVISION = "Unknown"
+
+    return GIT_REVISION
+
+
+if PYPI_VERSION is None:
+    PYPI_VERSION = git_version()
+
+
+## Standard setup
 
 this_directory = path.abspath(path.dirname(__file__))
 with io.open(path.join(this_directory, 'README.md'), encoding='utf-8') as f:
@@ -37,7 +70,7 @@ if __name__ == "__main__":
           author            = "Louis Moresi",
           author_email      = "louis.moresi@unimelb.edu.au",
           url               = "https://github.com/underworldcode/stripy",
-          version           = "0.7.0",
+          version           = PYPI_VERSION,
           description       = "Python interface to TRIPACK and STRIPACK fortran code for triangulation/interpolation in Cartesian coordinates and on a sphere",
           long_description  = long_description,
           long_description_content_type='text/markdown',
