### 2019-05-16 822de9d
Fixed duplicate and collinear points in square_border
