### 2019-05-16 11725c7
And the bugs found by adding the test !
