### 2019-05-20 #1410 @santisoler
Hi @VivianePons. The submitted code is robust and works as expected. The manuscript is very well written and the statement of need is very clear. The public functions are all documented inside the sources files and the repository has a lot of notebooks showing examples on how to use the code.

Most of the issues the authors are solving are related with the organization of the files inside the repository, pypi packaging, Docker container creation and improvement of automated tests in order to make them more robust.
I left some unchecked boxes in order to remember which aspects of the publication are not yet entirely improved. Although, most of them are being solved. Hope to get all of them checked off soon!

Would you please take a look at https://github.com/underworldcode/stripy/issues/14? We need some help regarding one of the references. Thanks!
