### 2019-05-22 b1c4aef
Updated README.md with recommended packages
