### 2019-05-22 underworldcode/stripy#24 @brmather
Thanks @santisoler - this has been addressed with b1c4aefcd82c7b7a669b79ebccbda94d167be4e6
