### 2019-05-28 afd1956
Create CONTRIBUTING.md
First pass at instructions for contributors.
