### 2019-05-28 underworldcode/stripy#30 @lmoresi
We added a contributions guide. This is modelled on other pieces of software from the underworld team and so is in our house style. We read the suggestions for inspiration too.
