### 2019-05-28 underworldcode/stripy#30 @santisoler
Great! :+1:

I think we could close this issue if you agree. Glad to see the repo looking so good!
