### 2019-05-28 underworldcode/stripy#14 @VivianePons
Yes, you can keep it without DOI if it does not exist
