### 2019-05-28 underworldcode/stripy#14 @santisoler
Thanks @VivianePons!
