### 2019-05-28 underworldcode/stripy#14 @santisoler
@lmoresi @brmather In light of this editorial feedback, we can safely close this issue if you agree.
