### 2019-05-29 ca6e651
Tests for expected functionality
