+++ tests/test_Triangulation.py
@@ -1,110 +1,125 @@
-import numpy as np
+import pytest
 import stripy
-
-from scipy import spatial
-from time import time
-
-try: range = xrange
-except: pass
-
-np.random.seed(0)
-x = np.random.random(500)
-y = np.random.random(500)
-
-# triangulation
-mesh = stripy.Triangulation(x, y)
-
+import numpy as np
 
 def test_derivative():
-    from scipy import interpolate
+    def analytic(x, y, k1, k2):
+        return np.cos(k1*x) * np.sin(k2*y)
 
-    # Create a field to test derivatives
-    x, y = mesh.x, mesh.y
-    Z = np.exp(-x**2 - y**2)
-    Zx = -2*x*Z
-    Zy = -2*y*Z
-    gradZ = np.hypot(Zx, Zy)
+    def analytic_dx(x, y, k1, k2):
+        return -k1 * np.sin(k1*x) * np.sin(k2*y)
 
-    # Stripy
-    t = time()
-    Zx1, Zy1 = mesh.gradient(Z, nit=10, tol=1e-10)
-    t1 = time() - t
-    gradZ1 = np.hypot(Zx1, Zy1)
-
-    # Spline
-    spl = interpolate.SmoothBivariateSpline(x, y, Z)
-    t = time()
-    Zx2 = spl.ev(x, y, dx=1)
-    Zy2 = spl.ev(x, y, dy=1)
-    t2 = time() - t
-    gradZ2 = np.hypot(Zx2, Zy2)
-
-    # Clough Tocher
-    # This one is most similar to what is used in stripy
-    t = time()
-    cti = interpolate.CloughTocher2DInterpolator(np.column_stack([x,y]),\
-                                                 Z, tol=1e-10, maxiter=20)
-    t3 = time() - t
-    Zx3 = cti.grad[:,:,0].ravel()
-    Zy3 = cti.grad[:,:,1].ravel()
-    gradZ3 = np.hypot(Zx3, Zy3)
-
-    res1 = ((gradZ1 - gradZ)**2).max()
-    res2 = ((gradZ2 - gradZ)**2).max()
-    res3 = ((gradZ3 - gradZ)**2).max()
-    print("squared error in first derivative\n  \
-           - stripy = {} took {}s\n  \
-           - spline = {} took {}s\n  \
-           - cloughtocher = {} took {}s".format(res1, t1, res2, t2, res3, t3))
+    def analytic_dy(x, y, k1, k2):
+        return k2 * np.cos(k1*x) * np.cos(k2*y)
 
 
-def test_interpolation():
-    from scipy import interpolate
+    extent = [0.0, 1.0, 0.0, 1.0]
+    mesh = stripy.cartesian_meshes.square_mesh(extent, spacingX=0.05, spacingY=0.05, refinement_levels=1)
+    
+    Z = analytic(mesh.x, mesh.y, 5.0, 2.0)
+    dZx, dZy = mesh.gradient(Z, nit=10, tol=1e-12)
 
-    x, y = mesh.x, mesh.y
-    Z = np.exp(-x**2 - y**2)
-    # We ensure interpolation points are within convex hull
-    xi = np.random.uniform(0.1, 0.9, 10)
-    yi = np.random.uniform(0.1, 0.9, 10)
+    dZx_analytic = analytic_dx(mesh.x, mesh.y, 5.0, 2.0)
+    dZy_analytic = analytic_dy(mesh.x, mesh.y, 5.0, 2.0)    
 
-    # Stripy
-    zn1 = mesh.interpolate_nearest(xi, yi, Z)
-    zl1, err = mesh.interpolate_linear(xi, yi, Z)
[…]
