### 2019-05-29 underworldcode/stripy#29 @brmather
This commit ca6e6511cba6e3cb29167b1464a25e8703b2496a tests the following functionality of the spherical and Cartesian triangulations:

- Derivatives `mesh.gradient(...)`
- Nearest neighbour interpolation `mesh.interpolate_linear(...)`
- Linear interpolation `mesh.interpolate_linear(...)`
- Cubic interpolation `mesh.interpolate_cubic(...)`
- Smoothing `mesh.smoothing(...)`

All tests are passing with `pytest` (which is a very useful tool by the way!) Let me know if there are any other tests you can think of, otherwise we can close this issue.

Cheers :+1:
