### 2019-05-29 underworldcode/stripy#29 @lmoresi
There will always be more tests - and we can check coverage with the relevant pytest plugin but the important thing is that we now have a reliable and extensible testing framework. We should run these in the final stages of the docker build to ensure we have at least some continuous checking. Then I would assert that we have done as asked and can close the issue.

Sent from my iPhone
