### 2019-05-29 01defbd
Incorporating pytest into Dockerfiles
