### 2019-05-29 underworldcode/stripy#29 @brmather
One can never have too many tests! Since 01defbdc81b383ce07306df0a378b953b21d588a `pytest` will run everything in the _tests_ directory.
