### 2019-05-31 39c6ed6
Initial look at compiled documentation with Sphinx
Requires more work than I anticipated, our docstrings largely
follow the numpy convention, but there are formatting issues
that mean sphinx cannot compile them correctly.
