### 2019-06-01 #1410 @lmoresi
**Response to reviews —  @VivianePons**

The review process has been thorough and very helpful. We felt that it might be useful to provide a brief summary of the action we have taken in response to the reviews as documented in the issue tracker:

  1. A number of the issues that were raised during the review were routine typos and minor bugs that we fixed as soon as they were brought to our attention. These include:
    - underworldcode/stripy#25
    - underworldcode/stripy#21
    - underworldcode/stripy#24
    - underworldcode/stripy#19

 2. Reviewer 2 (Renka) had some difficulty running the tests and so we incorporated `pytest` into our build process and also added the test suite to the dockerfile. Reviewer 1 (Soler) also commented on the lack of comprehensive unit testing and so we introduced a larger number of unit tests into the revised test suite. (underworldcode/stripy#21 , underworldcode/stripy#29 )

 3. Reviewer 1 commented on our use of the monorepo style of release and asked us to remove the `litho1pt0` module from the repository. We have done so. (underworldcode/stripy#26)

 4. Reviewer 1 commented on the structure of the repository: notebooks bundled in the source code and other idiosyncratic choices. We initially restructured the repository along the lines suggested by the reviewer but found that this did not permit us to bundle the documentation with the installation which is a reasonably common way to deal with a self-documenting module. We reached a compromise restructure with rev 1 which is closer to a standard python package repository but still allows bundling of the documentation in the form of notebooks. (underworldcode/stripy#27).

 5. Our approach to installation was to provide multiple pathways to working examples and that included `docker` images, links to `mybinder.org` deployment and installation via `pip`. Some of the issues raised by Reviewer 1 wrt the repository structure were associated with the need to have multiple branches to handle incompatible versions of (for example) dockerfiles for different pathways. This is something of a moving target and became easier recently. We were therefore able to remove some of the persistent branches and this is certainly an improvement in reducing fragility for future releases. This is discussed in underworldcode/stripy#17 in some length and also in underworldcode/stripy#16

  6. In underworldcode/stripy#17 there was also some discussion about release tags and semantic versioning. We partially addressed Reviewer 1's requirements here (we moved to a github-commit versioning for non-release installations) but have not dealt with this at the docker image / tagging level. This remains a manual process.

  7. We adopted Reviewer 1's suggestions for applying our licences via github so that they would be automatically recognised. We also added contribution guidelines as suggested. (underworldcode/stripy#15, underworldcode/stripy#30)

  8. Reviewer 1 asked us to consider using Sphinx to provide a web-based documentation stream for stripy. We agree that this is a nice suggestion but also felt that this should still be discretionary as it was not part of our original plan. We collectively closed that issue on the basis that we would consider this later - Reviewer 1 had offered his assistance and we felt that this would need to wait until the review was complete.   (underworldcode/stripy#28).

We think that we have addressed all the issued raised in the review process so far and hope that it can be drawn to a close. We would like to make a new release of the code that corresponds to to the version post review and would like to check if this is the standard operating procedure for JOSS and, if so, when it should be done.

Thanks for the opportunity to respond.
