### 2019-06-01 underworldcode/stripy#14 @lmoresi
I'm fine with that !
