### 2019-06-03 3c04d48
Bettering stripy docstrings
pdoc3 seems to be pretty good at generating documentation
sphinx is a pain and probably over-elaborate for what we need
