+++ stripy/__init__.py
@@ -1,5 +1,5 @@
 """
-Copyright 2017 Louis Moresi, Ben Mather
+Copyright 2017-2019 Louis Moresi, Ben Mather
 
 Stripy is free software: you can redistribute it and/or modify
 it under the terms of the GNU Lesser General Public License as published by
