+++ stripy/spherical.py
@@ -1,5 +1,5 @@
 """
-Copyright 2017 Louis Moresi, Ben Mather
+Copyright 2017-2019 Louis Moresi, Ben Mather
 
 This file is part of Stripy.
 
@@ -44,72 +44,70 @@ class sTriangulation(object):
     Define a Delaunay triangulation for given points on a sphere
     where lons and lats are 1D numpy arrays of equal length.
 
-    Algorithm
-    ---------
-     R. J. Renka (1997), Algorithm 772; STRIPACK: Delaunay triangulation
-     and Voronoi diagram on the surface of a sphere"
-     ACM Trans. Math. Softw., Vol. 23, No. 3, pp 416-434
-     doi:10.1145/275323.275329
+    Algorithm:
+        R. J. Renka (1997), Algorithm 772; STRIPACK: Delaunay triangulation
+        and Voronoi diagram on the surface of a sphere"
+        ACM Trans. Math. Softw., Vol. 23, No. 3, pp 416-434
+        doi:10.1145/275323.275329
 
-    Parameters
-    ----------
-     lons    : 1D array of longitudinal coordinates in radians
-     lats    : 1D array of latitudinal coordinates in radians
-     refinement_levels (int)
-             : refine the number of points in the triangulation
-               (see uniformly_refine_triangulation)
-     permute : (bool) randomises the order of lons and lats to improve
-               triangulation efficiency and eliminate colinearity
-               issues (see notes)
-     tree    : (bool) construct a cKDtree for efficient nearest-
-               neighbour lookup
+    Args:
+        lons : 1D array
+            longitudinal coordinates in radians
+        lats : 1D array
+            latitudinal coordinates in radians
+        refinement_levels : int
+            refine the number of points in the triangulation
+            (see uniformly_refine_triangulation)
+        permute : bool
+            randomises the order of lons and lats to improve
+            triangulation efficiency and eliminate colinearity issues (see notes)
+        tree : bool
+            construct a cKDtree for efficient nearest-neighbour lookup
 
-    Attributes
-    ----------
-     lons : array of floats, shape (n,)
-        stored longitudinal coordinates on a sphere
-     lats : array of floats, shape (n,)
-        stored latitudinal coordinates on a sphere
-     x : array of floats, shape (n,)
-        stored Cartesian x coordinates from input
-     y : array of floats, shape (n,)
-        stored Cartesian y coordinates from input
-     z : array of floats, shape (n,)
-        stored Cartesian z coordinates from input
-     simplices : array of ints, shape (nsimplex, 3)
-        indices of the points forming the simplices in the triangulation
-        points are ordered anticlockwise
-     lst : array of ints, shape (6n-12,)
-        nodal indices with lptr and lend, define the triangulation as a set of N
-        adjacency lists; counterclockwise-ordered sequences of neighboring nodes
-        such that the first and last neighbors of a boundary node are boundary
-        nodes (the first neighbor of an interior node is arbitrary).  In order to
-        distinguish between interior and boundary nodes, the last neighbor of
-        each boundary node is represented by the negative of its index.
-        The indices are 1-based (as in Fortran), not zero based (as in python).
-     lptr : array of ints, shape (6n-12),)
-        set of pointers in one-to-one correspondence with the elements of lst.
-        lst(lptr(i)) indexes the node which follows lst(i) in cyclical
-        counterclockwise order (the first neighbor follows the last neighbor).
-        The indices are 1-based (as in Fortran), not zero based (as in python).
-     lend : array of ints, shape (n,)
-        N pointers to adjacency lists.
-        lend(k) points to the last neighbor of node K.
-        lst(lend(K)) < 0 if and only if K is a boundary node.
[…]
