### 2019-06-03 underworldcode/stripy#28 @brmather
While this is independent of the formal review, I was interested to check out how this compiled documentation works. The workflow @santisoler suggests looks sound, but I encountered some issues with `sphinx`:

- `sphinx` pollutes the `docs` folder with makefiles, configuration files, etc. which I find messy.
- Markdown is not really supported and it is a pain to format everything to restructured text since everything is in markdown these days. The `recommonmark` package tries to fix this but does not *quite* cut it in my view.
- No matter how much I beat the docstrings to conform to numpy or google formatting, there were many instances where I simply could not get it to display correctly.
- Generally, I find the learning curve is quite high to make decent looking docs that render correctly.

Alternatives like `mkdocs` don't automatically compile docstrings, so that's out, but I stumbled across [`pdoc3`](https://pdoc3.github.io/pdoc/) which was a refreshingly pleasant experience. I only had to make very minimal changes to the docstrings to get it to format correctly and it requires no configuration files.

Just run:

```sh
pdoc3 --html -o docs/ stripy
```

and it compiles all the docstrings and shoves the HTML in your `docs` directory. I've been working off the [compiled-documentation](https://github.com/underworldcode/stripy/tree/compiled-documentation) branch if anyone want to take `pdoc3` for a spin. Remember to compile and install `stripy` from that branch first before attempting this as `pydoc3` looks up the installed version in the system/user directory to compile docstrings.
