### 2019-06-03 underworldcode/stripy#28 @lmoresi
The benefit seems to be that, like pytest, it is very lightweight and adds little overhead. Their motto is also one of the better ones out there.

L

Prof Louis Moresi

louis.moresi@unimelb.edu.au<mailto:louis.moresi@unimelb.edu.au>

(w) +61 3 8344 1217

(m) +61 4 0333 1413

(us) +1 505 349 4425

www.moresi.info<http://www.moresi.info/>

www.facebook.com/underworldcode<http://www.facebook.com/underworldcode>

@LouisMoresi<https://twitter.com/LouisMoresi>
