### 2019-06-03 underworldcode/stripy#28 @brmather
> Their motto is also one of the better ones out there.

:laughing:

@lmoresi can you set up the GitHub pages environment? I would do it but I don't have access to the settings panel for the repo...
