### 2019-06-03 a2f00f8
docs directory for documentation via pdoc3
