+++ docs/README.md
@@ -1,0 +1,3 @@
+### Documentation Pages
+
+Automatically built using the pdoc3 system.
