+++ docs/README.md
@@ -1,3 +1,13 @@
 ### Documentation Pages
 
 Automatically built using the pdoc3 system.
+
+#### Build the docs
+
+
+
+```bash
+
+pip install pdoc3
+pdoc --html stripy --force -o docs
+```
