+++ docs/README.md
@@ -4,10 +4,13 @@ Automatically built using the pdoc3 system.
 
 #### Build the docs
 
-
-
 ```bash
-
 pip install pdoc3
 pdoc --html stripy --force -o docs
+mv docs/stripy/*.html docs
 ```
+
+### Documentation site
+
+![www.underworldcode.org/stripy](http://www.underworldcode.org/stripy/)
+![underworldcode.github.io/stripy](http://underworldcode.github.io/stripy)
