### 2019-06-03 87cdd2f
Add a script to build the docs and clean up a bit (improve how we drive pdoc would be better). Also add links in docs strings to help with the built site.
