+++ stripy/__init__.py
@@ -12,6 +12,9 @@ GNU Lesser General Public License for more details.
 
 You should have received a copy of the GNU Lesser General Public License
 along with Stripy.  If not, see <http://www.gnu.org/licenses/>.
+
+Stripy source code is available ![on github](https://github.com/underworldcode/stripy)
+
 """
 
 from .spherical import sTriangulation
