+++ docs/build_pdocs.sh
@@ -1,6 +1,5 @@
 #! /usr/bin/env bash
 
-
 THIS_FILE=`stat -f ${BASH_SOURCE[0]} || readlink -f ${BASH_SOURCE[0]}`
 THIS_DIR=`dirname $THIS_FILE`
 echo "Building docs in $THIS_DIR"
