### 2019-06-03 underworldcode/stripy#28 @santisoler
> While this is independent of the formal review, I was interested to check out how this compiled documentation works.

Nice initiative! Documentation is a very important part of releasing software. Helping users to figure out how it works always makes things easy for them.

> The workflow @santisoler suggests looks sound, but I encountered some issues with `sphinx`:
>
>     * `sphinx` pollutes the `docs` folder with makefiles, configuration files, etc. which I find messy.

Your right, Sphinx builds the docs upon configuration files and a `Makefile`. Could be a little bit overwhelming at first, I agree.
Although, I think the `Makefile` is great when you need to tell a CI (continuous integration) to build the documentation: with a simple `make` command you ensure that the CI will perform the same task you are doing when building the documentation locally. Moreover, if some argument for the building has to be changed you just do it on the `Makefile`, leaving the CI configuration file unmodified.

>     * Markdown is not really supported and it is a pain to format everything to restructured text since everything is in markdown these days. The `recommonmark` package tries to fix this but does not _quite_ cut it in my view.

You're also right about this. It's a little bit annoying to use RST, Markdown is way simpler.
Nevertheless, you need to edit these files only once, or maybe when you make a new release, include a new contributor, etc. But I get it could be annoying.

>     * No matter how much I beat the docstrings to conform to numpy or google formatting, there were many instances where I simply could not get it to display correctly.

I always follow the [numpydoc docstring guide](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) when I struggle with these issues.
Nevertheless, the `stripy` docstrings are very well written and after a quick view I think they are in great agreement with the numpydoc docstring guide.
Maybe what's wrong is the single space before each parameter or returns. For example, I'll copy the `sTriangulation` docstring. I replaced these spaces with an underscore:
```
    Parameters
    ----------
    _lons    : 1D array of longitudinal coordinates in radians
    _lats    : 1D array of latitudinal coordinates in radians
    _refinement_levels (int)
    _        : refine the number of points in the triangulation
    _          (see uniformly_refine_triangulation)
    _permute : (bool) randomises the order of lons and lats to improve
    _          triangulation efficiency and eliminate colinearity
    _          issues (see notes)
    _tree    : (bool) construct a cKDtree for efficient nearest-
               neighbour lookup
```

I'm curious if that's the case, but don't worry if you don't want to test it. I can live with this doubt haha.

>     * Generally, I find the learning curve is quite high to make decent looking docs that render correctly.

Maybe you're right about this. That's why I haven't asked you to tackle down this issue in order to get the publication accepted.
From my experience, having an automated documentation that looks very nice afterwards worth the effort, but I agree that Sphinx could be a little bit difficult.
I'm glad it's not the only way to solve this.

> Alternatives like `mkdocs` don't automatically compile docstrings, so that's out, but I stumbled across [`pdoc3`](https://pdoc3.github.io/pdoc/) which was a refreshingly pleasant experience. I only had to make very minimal changes to the docstrings to get it to format correctly and it requires no configuration files.
>
> Just run:
>
> ```shell
> pdoc3 --html -o docs/ stripy
> ```
>
> and it compiles all the docstrings and shoves the HTML in your `docs` directory. I've been working off the [compiled-documentation](https://github.com/underworldcode/stripy/tree/compiled-documentation) branch if anyone want to take `pdoc3` for a spin. Remember to compile and install `stripy` from that branch first before attempting this as `pydoc3` looks up the installed version in the system/user directory to compile docstrings.

That's great! I have never used `pdoc` but seems a very nice and easy to use tool.
It's very well suited for `stripy` and the `pdoc` documentation (built by itself) looks very nice!
I can't wait to see how `stripy` documentation looks with `pdoc`.

Great job @brmather!
