### 2019-06-03 c0b9ca2
Add matplotlib & imageio recommended dependencies
