### 2019-06-03 088e0ed
Add instructions to run the tests
