### 2019-06-03 underworldcode/stripy#36 @santisoler
[REVIEW] Litho1pt0 importation is being tested
Part of https://github.com/openjournals/joss-reviews/issues/1410

Litho1pt0 has been moved away from Stripy repository, although its importation it's still being tested on [test/test_0_imports.py](https://github.com/underworldcode/stripy/blob/master/tests/test_0_imports.py).
I would suggest to remove it until `litho1pt0` is nicely packaged and could be installed though pypi.

What do you think?
