### 2019-06-03 underworldcode/stripy#36 @santisoler
Ok. I found that `litho1pt0` is still available on pypi.
So, maybe we should just add it to the `README.md`.
