### 2019-06-03 underworldcode/stripy#28 @lmoresi
The docs are live ... www.underworldcode.org/stripy (or underworldcode.github.io/stripy.

The one thing that would be good is a solution to the problem of embedding equations in the docstrings. It is possible with sphinx and it is somewhat possible with pdoc3 but it complains about the latex backslash escape characters which it is trying to interpret.

However, take a look. It seems to be at the right level for the project and, I must say, it does also make it helpful in keeping an eye on whether we are correctly contructing and populating the docstrings.
