### 2019-06-03 0c2e288
Removed traces of sphinx
