+++ docs/source/index.rst
@@ -1,35 +1,0 @@
-.. stripy documentation master file, created by
-   sphinx-quickstart on Fri May 31 11:40:45 2019.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to stripy's documentation!
-==================================
-
-Guide
-^^^^^
-
-.. toctree::
-   :maxdepth: 3
-   :caption: Contents:
-
-.. autoclass:: stripy.spherical.sTriangulation
-   :members:
-
-.. autofunction:: stripy.spherical.remove_duplicate_lonlat
-.. autofunction:: stripy.spherical.lonlat2xyz
-.. autofunction:: stripy.spherical.xyz2lonlat
-.. autofunction:: stripy.spherical.dxyz2dlonlat
-.. autofunction:: stripy.spherical.great_circle_Npoints
-.. autofunction:: stripy.spherical.angular_separation
-
-.. autoclass:: stripy.cartesian.Triangulation
-   :members:
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
