### 2019-06-03 underworldcode/stripy#28 @brmather
Ah, be sure to compile off the "compiled-documentation" branch. I made a few minor adjustment to the docstrings to get them to render correctly (see pull request #37).
