### 2019-06-04 7679e41
Updated documentation
