### 2019-06-04 underworldcode/stripy#28 @brmather
OK the [docs](http://www.underworldcode.org/stripy/) are looking *very, very* good!

> The one thing that would be good is a solution to the problem of embedding equations in the docstrings. It is possible with sphinx and it is somewhat possible with pdoc3 but it complains about the latex backslash escape characters which it is trying to interpret.

I'm not sure how to embed maths expressions, however, the backslash shouldn't be a problem if you put *r* in front of the docstring, right? If not, it might be worth raising an [issue ](https://github.com/pdoc3/pdoc/issues) on their GitHub since `pdoc` is under active development.
