### 2019-06-04 125a284
Update docs with some equations. Is this too confusing when reading the docstrings manually or in a notebook ?
