### 2019-06-04 underworldcode/stripy#28 @lmoresi
https://github.com/lmoresi/pdoc - I made a quick fork and added some mathjax support and python markdown support. The issue of escaping still arises which can be avoided by escaping the escapes or by marking the docstring as raw. There are some deprecated features in the markdown math extension but ... details.

It's not obvious which is the right choice but I am not that keen on the raw marker as it looks horrible. I am also not sure if it would have side effects to coerce all docstrings to be interpreted as raw when pdoc launches.

Have a look though, because there are quite a few places where it would be useful to introduce mathematical notation (e.g. in the derivatives). We can ask the pdoccers about a PR when we have a working example posted ... see below for a first example

http://www.underworldcode.org/stripy/cartesian.html#stripy.cartesian.Triangulation.gradient
