### 2019-06-04 97d4815
More MathJax expressions in cartesian.py
