+++ stripy/cartesian.py
@@ -94,14 +94,14 @@ class Triangulation(object):
 
     Notes:
         Provided the nodes are randomly ordered, the algorithm
-        has expected time complexity O(N*log(N)) for most nodal
+        has expected time complexity \\(O(N*log(N)\\) for most nodal
         distributions.  Note, however, that the complexity may be
         as high as \\(O(N^2)\\) if, for example, the nodes are ordered
         on increasing x.
 
-        If permute=True, x and y are randomised on input before
+        If `permute=True`, x and y are randomised on input before
         they are triangulated. The distribution of triangles will
-        differ between setting permute=True and permute=False,
+        differ between setting `permute=True` and `permute=False`,
         however, the node ordering will remain identical.
     """
     def __init__(self, x, y, refinement_levels=0, permute=False, tree=False):
@@ -263,7 +263,7 @@ class Triangulation(object):
         The method consists of minimizing a quadratic functional Q(G) over
         gradient vectors (in x and y directions), where Q is an approximation
         to the linearized curvature over the triangulation of a C-1 bivariate
-        function F(x,y) which interpolates the nodal values and gradients.
+        function \\(F(x,y)\\) which interpolates the nodal values and gradients.
 
         Args:
             f : array of floats, shape (n,)
@@ -321,8 +321,8 @@ class Triangulation(object):
 
         This routine employs a local method, in which values depend only on nearby
         data points, to compute an estimated gradient at a node.
-
-        gradient_local() is more efficient than gradient() only if it is unnecessary
+        
+        `gradient_local()` is more efficient than `gradient()` only if it is unnecessary
         to compute gradients at all of the nodes. Both routines have similar accuracy.
         """
         if f.size != self.npoints:
@@ -339,7 +339,7 @@ class Triangulation(object):
 
     def smoothing(self, f, w, sm, smtol, gstol):
         """
-        Smooths a surface f by choosing nodal function values and gradients to
+        Smooths a surface `f` by choosing nodal function values and gradients to
         minimize the linearized curvature of F subject to a bound on the
         deviation from the data values. This is more appropriate than interpolation
         when significant errors are present in the data.
@@ -348,24 +348,24 @@ class Triangulation(object):
             f : array of floats, shape (n,)
                 field to apply smoothing on
             w : array of floats, shape (n,)
-                weights associated with data value in f
-                w[i] = 1/sigma_f^2 is a good rule of thumb.
+                weights associated with data value in `f`
+                `w[i] = 1/sigma_f**2` is a good rule of thumb.
             sm : float
                 positive parameter specifying an upper bound on Q2(f).
-                generally n-sqrt(2n) <= sm <= n+sqrt(2n)
-            smtol : float
+                generally `n-sqrt(2n) <= sm <= n+sqrt(2n)`
+            smtol : float [0,1]
                 specifies relative error in satisfying the constraint
-                sm(1-smtol) <= Q2 <= sm(1+smtol) between 0 and 1.
+                `sm(1-smtol) <= Q2 <= sm(1+smtol)` between 0 and 1.
             gstol : float
                 tolerance for convergence.
-                gstol = 0.05*mean(sigma_f)^2 is a good rule of thumb.
+                `gstol = 0.05*mean(sigma_f)**2` is a good rule of thumb.
 
         Returns:
             f_smooth : array of floats, shape (n,)
                 smoothed version of f
             derivatives : tuple of floats, shape (n,3)
-                (dfdx, dfdy) first derivatives of f_smooth in the
-                x and y directions
+                \\( \\frac{df}{dx} , \\frac{df}{dy} \\) first derivatives
+                of `f_smooth` in the x and y directions
         """
         if f.size != self.npoints or f.size != w.size:
[…]
