### 2019-06-04 underworldcode/stripy#28 @brmather
This seems pretty good. It supports fractions and symbols e.g. [`interpolate_cubic()`](http://www.underworldcode.org/stripy/cartesian.html#stripy.cartesian.Triangulation.interpolate_cubic). The MathJax(ed) docs look decent for `cartesian.py` and it's only a matter of time before `spherical.py` looks as good!
