### 2019-06-04 af1495a
Added Travis CI configuration file
