+++ .travis.yml
@@ -1,0 +1,30 @@
+dist: bionic
+language: python
+python:
+  - "3.5"
+  - "3.6"
+  - "3.7"
+
+
+install:
+  - pip3 install numpy scipy pytest git+https://github.com/lmoresi/pdoc.git
+  - pip3 install -e .
+
+script:
+  # generate the docs
+  - mkdir docs
+  - cd docs 
+  - pdoc --html -o . --force stripy
+  - mv pycurious/*.html .
+  - rm -rf pycurious/
+  - cd ../
+
+deploy:
+  provider: pages
+  skip_cleanup: true
+  github_token: $GITHUB_TOKEN  # Set in the settings page of your repository, as a secure variable
+  keep_history: true
+  on:
+    branch: master
+  verbose: true
+  local_dir: docs/
