### 2019-06-04 underworldcode/stripy#28 @santisoler
:tada: Nice job! :tada:

The documentation looks really good!
I liked how pdoc created a very beautiful documentation with a very low learning curve.
Congratulations! :muscle:

One minor thought.
Have you compiled the documentation locally and then pushed it to the repo?
This approach is very quick but might be annoying in the future: each time you make some change on the code, you need to remember to update the documentation.
Another approach would be to configure a CI (Travis CI is free for public Github repositories) to compile the documentation and push it to a dedicated branch, e.g. `gh-pages`. That branch will be only handled by the CI, which will update it when a new commit is pushed to `master`.
From my experience, working with CIs makes maintenance a lot easier and allow us to focus more on the most important part of the repository (design, code, how docstrings are written, etc).

The [Travis CI documentation](https://docs.travis-ci.com/) is very easy to understand. To configure it you'll just need to write a `.travis.yml` file, and on this specific case you may need some scripts for deploying the documentation into the `gh-pages` branch.
@leouieda wrote some continuous integration scripts on [fatiando/continuous-integration](https://github.com/fatiando/continuous-integration). Some of them are [specifically for Travis](https://github.com/fatiando/continuous-integration/tree/master/travis).
I always refer to these scripts when I need to configure Travis and/or deploy from it.

Just want to make your life easier with this advice.
I think I'll spend a few more minutes appreciating the new Stripy documentation! :relaxed:
