### 2019-06-04 underworldcode/stripy#28 @santisoler
I've just seen you started configuring Travis a few hours ago! That's great!
Hope my previous message helps.
