### 2019-06-04 underworldcode/stripy#28 @brmather
Yes @santisoler I had the same thought! Over time it would get very tedious to manually compile the documentation. @lmoresi I don't think the repo is configured to let me manage `stripy` on Travis ([travis-ci.org](https://travis-ci.org) or [travis-ci.com](https://travis-ci.com/)). In any case, I tried it out my [`pycurious`](https://github.com/brmather/pycurious) package and it works really well. I've copied the `.travis.yml` file to `stripy` so setting it up should be relatively straightforward. Here are the steps I took:

1. Setup `stripy` in Travis CI with a [personal access token](https://github.com/settings/tokens)
2. Create a *gh-pages* branch for Travis to push the documentation. You can also configure it to go in the `docs` folder, but I find this messy because Travis ends up committing to the master branch, and the HTML pages inside `docs` don't render in GitHub anyway - so I'd prefer if that folder was deleted.
3. Configure the `.travis.yml` setup (skip this because I've already done it)
4. Grab a gin & tonic because that's it! :cocktail:

I haven't configured Travis to run pytest on all the test scripts because some of those optional packages that are imported in `test_0_imports.py` would be tricky to setup inside the environment (I'm looking at you `lavavu`, `gdal` and `cartopy`! :eyes: ) So we should make a decision whether those tests ought to stay...
