### 2019-06-04 underworldcode/stripy#28 @brmather
@santisoler I just looked those Travis CI scripts on fatiando - the one on deploying to PyPI is super useful! Thanks :+1:
