### 2019-06-04 underworldcode/stripy#28 @santisoler
Glad to be helpful!

> In any case, I tried it out my pycurious package and it works really well. I've copied the .travis.yml file to stripy so setting it up should be relatively straightforward.

Nice package you have! I'll leave a comment about it on its repo.

> 1. Setup `stripy` in Travis CI with a [personal access token](https://github.com/settings/tokens)
>
> 1. Create a _gh-pages_ branch for Travis to push the documentation. You can also configure it to go in the `docs` folder, but I find this messy because Travis ends up committing to the master branch, and the HTML pages inside `docs` don't render in GitHub anyway - so I'd prefer if that folder was deleted.
>
> 1. Configure the `.travis.yml` setup (skip this because I've already done it)
>
> 1. Grab a gin & tonic because that's it! :cocktail:

That's great! Specially the last step haha!
Be careful with the personal access token! You **must encrypt** it and prevent Travis from dumping it into the Job Log (see [this line](https://github.com/fatiando/continuous-integration/blob/master/travis/deploy-gh-pages.sh#L32)).
If you have the slightly doubt that the access token has been compromised (unencrypted published or dumped by Travis on the logs), you should delete it and create a new one.

> I haven't configured Travis to run pytest on all the test scripts because some of those optional packages that are imported in `test_0_imports.py` would be tricky to setup inside the environment (I'm looking at you `lavavu`, `gdal` and `cartopy`! :eyes: ) So we should make a decision whether those tests ought to stay...

I normally use `conda` for managing python packages and dependencies.
I had no problem installing `gdal` or `cartopy` from the `conda-forge`, so that may help.
An easy way to get `conda` installed in Travis is through the [`setup-miniconda.sh`](https://github.com/fatiando/continuous-integration/blob/master/travis/setup-miniconda.sh) script.
You could create an `environment.yml` detailing the dependencies for Stripy and telling `conda` to create an environment with those dependencies (`conda` will create the environment, download and install all the dependencies).
If no package is available on `conda-forge` you could alternatively install it through `pip`.
Buuuut, I haven't been able to install `lavavu` from `pip`.

So, I think is not to hard to make Travis run pytest, but you will encounter some issues with `lavavu`.

I'm really happy with the shape Stripy is taking. Sorry to be repetitive, but I'm glad with all the effort you are putting on this. You earned that gin & tonic! Cheers!
