### 2019-06-04 underworldcode/stripy#28 @lmoresi
I’d suggest we consider docker here. Should be reasonable to have a stable base image and build against that. This would be advantageous given our reliance on docker for nearly every other workflow - and because of the complexity of our dependencies.

This would be a good thing for other packages and deployments (e.g. of courses) that contain underworld dependencies where the setup is fragile and murderously difficult.

Conda is not a friend to me. I have all sorts of scars from trying to make it do what I want in the broader environment.

Prof Louis Moresi

louis.moresi@unimelb.edu.au<mailto:louis.moresi@unimelb.edu.au>

(w) +61 3 8344 1217

(m) +61 4 0333 1413

(us) +1 505 349 4425

www.moresi.info<http://www.moresi.info/>

www.facebook.com/underworldcode<http://www.facebook.com/underworldcode>

@LouisMoresi<https://twitter.com/LouisMoresi>
