### 2019-06-04 underworldcode/stripy#28 @brmather
Yes fair enough, we already run the tests in the Docker build using `pytest` so no need to double up with Travis CI.
