### 2019-06-05 underworldcode/stripy#28 @santisoler
Configuring Travis to run the tests from within your Docker image would be nice, specially if that's your best recommended way to run Stripy.

My conda proposal was just to run the tests for the non-Docker installation. You could either use the standard Python distribution included on Travis CI VMs, but @brmather exposed some issues regarding a few dependencies (`gdal`, `cartopy`, `lavavu`). That's why I offered the conda alternative.

In summary, I agree with @lmoresi: running the tests from within your Docker image should be implemented. Then you can discuss if you want to implement the automated tests on any other platform.
