### 2019-06-05 underworldcode/stripy#28 @lmoresi
Thanks - I had in mind the idea that we would run the tests in a docker environment with all the necessary dependencies which seems to be a documented workflow for travis. This replaces the build-or-die that we currently have with the automated docker build running the tests as the last command in the dockerfile and then flagging a fail-to-build on dockerhub !

Prof Louis Moresi

louis.moresi@unimelb.edu.au<mailto:louis.moresi@unimelb.edu.au>

(w) +61 3 8344 1217

(m) +61 4 0333 1413

(us) +1 505 349 4425

www.moresi.info<http://www.moresi.info/>

www.facebook.com/underworldcode<http://www.facebook.com/underworldcode>

@LouisMoresi<https://twitter.com/LouisMoresi>
