### 2019-06-05 #1410 @santisoler
Hi @VivianePons.

I don't think I can improve @lmoresi summary. But I can add a few things.

I consider the paper and the repository have gone through a great improvement in comparison with the first submitted version.
Regarding JOSS's requirements, I think they are all almost satisfied. I open a small issue (https://github.com/underworldcode/stripy/issues/36) and two Pull Requests (https://github.com/underworldcode/stripy/pull/34 and https://github.com/underworldcode/stripy/pull/35) in order to add some instructions to the `README.md`. That's why I haven't checked off the **Installation instructions** box.
The other box left to check off is the one related with the **Version**. I'm awaiting for the authors to make a new release whenever they feel ready.

I've also raised the possibility to implement automated test through `pytest` and compiled documentation through Sphinx. Although I stated that *solving these issues was not mandatory* to accept the publication, the authors implemented the automated tests with `pytest` and published a compiled documentation to GitHub Pages using [`pdoc`](https://pdoc3.github.io/pdoc/) instead of Sphinx. They found `pdoc` to be better suited than Sphinx due to the not so large size of their repository and its ease of use. I didn't even know `pdoc` existence, but I found it out to be a very nice alternative to Sphinx for small to medium projects.
I think it would be nice to have it into account for other JOSS papers that wants to quickly implement nice looking documentation.

In summary, when the `README.md` is updated accordingly to the issues I raised and a new release is created, I'll be ready to accept the paper for publication.
Nevertheless, I feel the authors are very excited by implementing a Continuous Integration to automate the tests and the documentation build. Maybe would be nice to wait for them to be ready. That's up to them.

Finally I want to thanks the authors for the effort and responsiveness to each of one of the issues and comments I opened. They were always very respectful and created a nice atmosphere to discuss possible solutions to those issues. Besides they were always open to take into account each one of my advises, even if that implied to dive into new software or tools. I know the effort that takes reading through documentation for the first time and get used to a new tool, so I really appreciate that.
Also, I want to thanks the authors for the patience with my use of English. It's not my first language and I apologize if it led to any misunderstanding.
I had a very nice experience reviewing this paper. I really liked the review process of JOSS and would be glad to do it again if needed.
Thanks you @VivianePons and every one behind JOSS for all your work.
