### 2019-06-06 eea3be0
Better tests for functionality
... that should be agnostic to rounding errors
