+++ tests/test_sTriangulation.py
@@ -2,47 +2,29 @@ import pytest
 import stripy
 import numpy as np
 
-def test_derivative():
-    def analytic(lons, lats, k1, k2):
-        return np.cos(k1*lons) * np.sin(k2*lats)
-
-    def analytic_ddlon(lons, lats, k1, k2):
-        return -k1 * np.sin(k1*lons) * np.sin(k2*lats) / np.cos(lats)
-
-    def analytic_ddlat(lons, lats, k1, k2):
-        return k2 * np.cos(k1*lons) * np.cos(k2*lats)
-
-
-    mesh = stripy.spherical_meshes.icosahedral_mesh(include_face_points=True, refinement_levels=3)
-    
-    Z = analytic(mesh.lons, mesh.lats, 5.0, 2.0)
-    dZdlon, dZdlat = mesh.gradient_lonlat(Z, nit=10, tol=1e-12)
-
-    dZdlon_analytic = analytic_ddlon(mesh.lons, mesh.lats, 5.0, 2.0)
-    dZdlat_analytic = analytic_ddlat(mesh.lons, mesh.lats, 5.0, 2.0)    
-
-    if np.isclose(dZdlon, dZdlon_analytic, 1.,1.).all() and \
-       np.isclose(dZdlat, dZdlat_analytic, 1.,1.).all():
-        print("PASS! (Derivatives)")
-    else:
-        assert False, "FAIL! (Derivatives)"
-
-
 def test_nearest_nd_interpolation():
+
+    p0 = 0.0
+    p2 = np.pi/2
+    p4 = np.pi/4
     
-    coords = np.array([[0.0, 0.0], \
-                       [0.0, 1.0], \
-                       [1.0, 0.0], \
-                       [1.0, 1.0]])
+    coords = np.array([[p0 , -p2], \
+                       [-p2,  p0], \
+                       [p0 ,  p2], \
+                       [p2 ,  p0]])
 
     lons, lats = coords[:,0], coords[:,1]
     mesh = stripy.sTriangulation(lons, lats)
 
-    Z = mesh.lons + mesh.lats
+    Z = np.linspace(-p2, p2, mesh.npoints)
 
-    Zi, ierr = mesh.interpolate_nearest(0.9, 0.9, Z)
+    ilons = lons[0] + 0.001
+    ilats = lats[0] + 0.001
+
+    Zi, ierr = mesh.interpolate_nearest(ilons, ilats, Z)
     
-    if Zi == 2.0:
+    # check if we return the nearest numbers
+    if Zi == Z[0]:
         print("PASS! (Interpolation - nearest neighbour)")
     else:
         assert False, "FAIL! (Interpolation - nearest neighbour)"
@@ -50,19 +32,36 @@ def test_nearest_nd_interpolation():
 
 def test_linear_interpolation():
 
-    coords = np.array([[0.0, 0.0], \
-                       [0.0, 1.0], \
-                       [1.0, 0.0], \
-                       [1.0, 1.0]])
+    p0 = 0.0
+    p2 = np.pi/2
+    p4 = np.pi/4
+
+    coords = np.array([[p0 , -p2], \
+                       [-p2,  p0], \
+                       [p0 ,  p2], \
+                       [p2 ,  p0], \
+                       [p0 ,  p0]])
 
     lons, lats = coords[:,0], coords[:,1]
     mesh = stripy.sTriangulation(lons, lats)
 
-    Z = mesh.lons + mesh.lats
+    Z = mesh.lons
 
-    Zi, ierr = mesh.interpolate_linear(0.5, 0.5, Z)
+    npts = 5
+    ilons = np.linspace(-p2, p2, npts)
[…]
