+++ tests/test_Triangulation.py
@@ -2,33 +2,6 @@ import pytest
 import stripy
 import numpy as np
 
-def test_derivative():
-    def analytic(x, y, k1, k2):
-        return np.cos(k1*x) * np.sin(k2*y)
-
-    def analytic_dx(x, y, k1, k2):
-        return -k1 * np.sin(k1*x) * np.sin(k2*y)
-
-    def analytic_dy(x, y, k1, k2):
-        return k2 * np.cos(k1*x) * np.cos(k2*y)
-
-
-    extent = [0.0, 1.0, 0.0, 1.0]
-    mesh = stripy.cartesian_meshes.square_mesh(extent, spacingX=0.05, spacingY=0.05, refinement_levels=1)
-    
-    Z = analytic(mesh.x, mesh.y, 5.0, 2.0)
-    dZx, dZy = mesh.gradient(Z, nit=10, tol=1e-12)
-
-    dZx_analytic = analytic_dx(mesh.x, mesh.y, 5.0, 2.0)
-    dZy_analytic = analytic_dy(mesh.x, mesh.y, 5.0, 2.0)    
-
-    if np.isclose(dZx, dZx_analytic, 1.,1.).all() and \
-       np.isclose(dZy, dZy_analytic, 1.,1.).all():
-        print("PASS! (Derivatives)")
-    else:
-        assert False, "FAIL! (Derivatives)"
-
-
 def test_nearest_nd_interpolation():
     
     coords = np.array([[0.0, 0.0], \
@@ -39,11 +12,14 @@ def test_nearest_nd_interpolation():
     x, y = coords[:,0], coords[:,1]
     mesh = stripy.Triangulation(x, y)
 
-    Z = mesh.x + mesh.y
+    Z = np.linspace(0.0, 10.0, mesh.npoints)
 
-    Zi, ierr = mesh.interpolate_nearest(0.9, 0.9, Z)
+    ix = x[0] + 0.001
+    iy = y[0] + 0.001
+
+    Zi, ierr = mesh.interpolate_nearest(ix, iy, Z)
     
-    if Zi == 2.0:
+    if Zi == Z[0]:
         print("PASS! (Interpolation - nearest neighbour)")
     else:
         assert False, "FAIL! (Interpolation - nearest neighbour)"
@@ -54,16 +30,28 @@ def test_linear_interpolation():
     coords = np.array([[0.0, 0.0], \
                        [0.0, 1.0], \
                        [1.0, 0.0], \
-                       [1.0, 1.0]])
+                       [1.0, 1.0], \
+                       [0.5, 0.5]])
 
     x, y = coords[:,0], coords[:,1]
     mesh = stripy.Triangulation(x, y)
 
-    Z = mesh.x + mesh.y
+    Z = mesh.x
 
-    Zi, ierr = mesh.interpolate_linear(0.5, 0.5, Z)
+    npts = 5
+    ix = np.linspace(0.0, 1.0, npts)
+    iy = np.zeros(npts)
 
-    if np.isclose(Zi, 1.0, 0.001):
+    Zi, ierr = mesh.interpolate_linear(ix, iy, Z)
+
+    # this should be true
+    # but machine precision may differ so we don't test it
+    # print((Zi == ix).all())
+
+    bounded = Zi[0] == ix[0] and Zi[-1] == ix[-1]
+    ascending = ( np.diff(Zi) > 0 ).all()
+
+    if bounded and ascending:
         print("PASS! (Interpolation - linear")
     else:
         assert False, "FAIL! (Interpolation - linear)"
@@ -84,14 +72,66 @@ def test_cubic_interpolation():
     x, y = coords[:,0], coords[:,1]
     mesh = stripy.Triangulation(x, y)
 
-    Z = mesh.x + mesh.y
+    Z = mesh.x**2
 
[…]
