### 2019-06-06 ad6eaec
Add a link to pytest on pypi
Helps encourage people to install pytest and run in `stripy`
