### 2019-06-06 bfdd5dc
Add litho1pt0 as optional dependency
