### 2019-06-06 underworldcode/stripy#36 @brmather
Alright, I've just added it with your pull request #34
