### 2019-06-06 14f4300
Implement testing on multiple architectures
Run `pytest` on Linux and OSX. I remove the optional dependencies
from the `test_0_imports.py` package.
