+++ .travis.yml
@@ -1,22 +1,43 @@
-dist: bionic
 language: python
-python:
-  - "3.5"
-  - "3.6"
-  - "3.7"
 
+# define global environment variables
+env:
+  global:
+    -DEPLOY_DOCS=false
+    -DEPLOY_PYPI=false
+
+matrix:
+  include:
+    - os: linux
+      dist: trusty
+      python: 3.5
+      env:
+        - DEPLOY_DOCS=true
+        - DEPLOY_PYPI=true
+    - os: linux
+      dist: bionic
+      python: 3.7
+    - os: osx
+      osx_image: xcode10.2
+      language: generic
+
+before_install:
+  - if [ "$TRAVIS_OS_NAME" = "osx" ]; then brew update ; fi
+  - if [ "$TRAVIS_OS_NAME" = "osx" ]; then brew install python3 ; fi
 
 install:
-  - pip3 install numpy scipy pytest git+https://github.com/lmoresi/pdoc.git
-  - pip3 install -e .
+  - python3 -m pip install Cython numpy scipy pytest git+https://github.com/lmoresi/pdoc.git
+  - python3 -m pip install -e .
 
 script:
+  # run pytest
+  - pytest tests/
   # generate the docs
   - mkdir docs
   - cd docs 
   - pdoc --html -o . --force stripy
-  - mv pycurious/*.html .
-  - rm -rf pycurious/
+  - mv stripy/*.html .
+  - rm -rf stripy/
   - cd ../
 
 deploy:
@@ -26,5 +47,6 @@ deploy:
   keep_history: true
   on:
     branch: master
+    condition: '$DEPLOY_DOCS == "true"'
   verbose: true
   local_dir: docs/
