### 2019-06-06 underworldcode/stripy#28 @brmather
In light of potential issues on other platforms #38 - I think it's we should use Travis CI to test on Linux and OSX. I've created a pull request #40 to update `.travis.yml` configuration file. Amongst other changes, I have removed these optional dependency tests from `test_0_imports.py`:

```python
def test_documentation_dependencies():
    import matplotlib
    import cartopy
    import imageio
    import lavavu
    import pyproj

def test_litho1pt0():
    import litho1pt0
```

A lot of users might not want to install the optional dependencies and just want to install stripy to their system directory, in which case I think we ought to integrate testing on multiple architectures. If they want to run the examples, then they can run the Docker image or Binder. The installation of all the optional dependencies is not a trivial feat!
