### 2019-06-07 cde0e79
Cartesian interpolation / extrapolation refactoring
Properly return interpolation / extrapolation result for
`interpolate_nearest()` method in Cartesian module.
