+++ stripy/cartesian.py
@@ -416,8 +416,7 @@ class Triangulation(object):
         """
 
         if order == 0:
-            zierr = np.zeros_like(xi, dtype=np.int)
-            return self.interpolate_nearest(xi, yi, zdata), zierr
+            return self.interpolate_nearest(xi, yi, zdata)
         elif order == 1:
             return self.interpolate_linear(xi, yi, zdata)
         elif order == 3:
@@ -440,16 +439,40 @@ class Triangulation(object):
         Returns:
             zi : float / array of floats, shape (l,)
                 nearest-neighbour interpolated value(s) of (xi,yi)
+            err : int / array of ints, shape (l,)
+                whether interpolation (0), extrapolation (1) or error (other)
         """
         if zdata.size != self.npoints:
             raise ValueError('zdata should be same size as mesh')
 
+        xi = np.atleast_1d(xi)
+        yi = np.atleast_1d(yi)
+
+        size = xi.size
+
         zdata = self._shuffle_field(zdata)
 
-        ist = np.ones_like(xi, dtype=np.int32)
+        zierr = np.zeros(size, dtype=np.int32)
+        ist = np.ones(size, dtype=np.int32)
         ist, dist = _tripack.nearnds(xi, yi, ist, self._x, self._y,
                                      self.lst, self.lptr, self.lend)
-        return zdata[ist - 1], ist*0
+
+        # check if interpolation or extrapolation
+        hull_idx = self.convex_hull()
+        hull_pts = self.points[hull_idx]
+        hull_x = hull_pts[:,0]
+        hull_y = hull_pts[:,1]
+
+        for i in range(0, zierr.size):
+            vector_det = (hull_x[1:] - hull_x[:-1])*(yi[i] - hull_y[:-1]) - \
+                         (hull_y[1:] - hull_y[:-1])*(xi[i] - hull_x[:-1])
+            
+            # if vector_det > 0: within convex hull
+            # if vector_det = 0: on top of convex hull
+            # if vector_det < 0: outside convex hull
+            zierr[i] = (vector_det < 0).any()
+
+        return np.squeeze(zdata[ist - 1]), zierr
 
 
     def interpolate_linear(self, xi, yi, zdata):
@@ -523,7 +546,7 @@ class Triangulation(object):
             err : int / array of ints, shape (l,)
                 whether interpolation (0), extrapolation (1) or error (other)
             dzx, dzy (optional) : float, array of floats, shape(l,)
-                first partial derivatives \\( \\frac{df}{dx} , \\frac{df}{dy} \\)
+                first partial derivatives \\( \\partial f \\partial x , \\partial f \\partial y \\)
                 at (xi,yi)
         """
 
