### 2019-06-10 #1410 @santisoler
All the issues I open are already solved, so I checked off every box but the **Version** one.
Once the authors make a new release, I consider the submitted paper will be ready for publication.
