### 2019-06-11 #1410 @VivianePons
Perfect! The editorial process will include releasing a new version anyway and I'll update it here.

I consider the paper to be accepted, good job everyone! It was great to see so much interactions between reviewers and authors.
