### 2019-06-11 #1410 @VivianePons
@lmoresi could you:

 * release the new updated version
 * create Zenodo archive

(question: is the "from branch master" still necessary to generate the pdf?)
