### 2019-06-11 #1410 @arfon
> (question: is the "from branch master" still necessary to generate the pdf?)

Whedon by default uses `master` so this is redundant.
