### 2019-06-11 #1410 @lmoresi
I explicitly stated “master” last time simply to stop anyone (likely myself) pasting the previous command that used the “paper” branch.
Redundant in every computational sense other than to account for human fallibility.

L

Prof Louis Moresi

louis.moresi@unimelb.edu.au<mailto:louis.moresi@unimelb.edu.au>

(w) +61 3 8344 1217

(m) +61 4 0333 1413

(us) +1 505 349 4425

www.moresi.info<http://www.moresi.info/>

www.facebook.com/underworldcode<http://www.facebook.com/underworldcode>

@LouisMoresi<https://twitter.com/LouisMoresi>
