### 2019-06-11 645d6eb
I presume this is supposed to be stripy not pycurious
