### 2019-06-11 c627d3f
Switch up to version 1.0.0
