### 2019-06-11 2531258
Test failing due to "colinear" points in soccerball mesh.
I think this actually refers to the first 3 points all lying on a great 
circle. I have set permute to true by default for this mesh and also 
updated the feedback message.
