### 2019-06-11 bccd145
I had to increment to 1.0.1 due to an upload problem with 1.0.0 on PYPI
