### 2019-06-11 #1410 @lmoresi
@VivianePons

I have made the release (it's 1.0.1 on both PYPI and Github) and it is automatically propagated to zenodo at [doi.org/10.5281/zenodo.3243511](https://doi.org/10.5281/zenodo.3243511)

whedon can automatically build from the default branch now.

Thanks for all the help. That was a very useful / interesting experience and we will be happy to publish with JOSS again.

> @lmoresi could you:
> * release the new updated version
> * create Zenodo archive
