### 2019-06-13 #1410 @VivianePons
Congratulations! Your paper has been officially accepted to JOSS. I leave it to @openjournals/joss-eics for finalization
