### 2019-06-13 8bbfe03
Failing on clean docker build fortran path problem
