### 2019-06-13 fb7fbc4
There is an issue with missing functions on debian if the include.f is not in the path but it cannot just be added to the extension. For now, add all files into the same source tree.
