### 2019-06-13 #1410 @arfon
@lmoresi - could you merge this PR, it fixes the position of the figure caption: https://github.com/underworldcode/stripy/pull/41
