### 2019-06-14 #1410 @QuLogic
The Software repository link on the final article is 404. It should probably point to the repo, not a subdirectory.
