### 2019-06-14 #1410 @lmoresi
I have added a ‘paper’ tag to the repo that snapshots the current / published state so that the link no longer fails.

However, this is a snapshot of the published release and I am not sure that this is what is intended by that link.

Prof Louis Moresi

louis.moresi@unimelb.edu.au<mailto:louis.moresi@unimelb.edu.au>

(w) +61 3 8344 1217

(m) +61 4 0333 1413

(us) +1 505 349 4425

www.moresi.info<http://www.moresi.info/>

www.facebook.com/underworldcode<http://www.facebook.com/underworldcode>

@LouisMoresi<https://twitter.com/LouisMoresi>
